% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/migrate-hackery.R
\name{acl_add_group}
\alias{acl_add_group}
\alias{acl_remove_group}
\title{ACL Add Group}
\usage{
acl_add_group(content, group_guid, role)

acl_remove_group(content, group_guid)
}
\arguments{
\item{content}{The R6 Content object}

\item{group_guid}{The group's GUID}

\item{role}{One of "owner" or "viewer"}
}
\value{
The R6 content object (for piping)
}
\description{
Add a group_guid to the content as an owner or viewer
}
\keyword{internal}
