% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/migrate-hackery.R
\name{acl_add_user}
\alias{acl_add_user}
\alias{acl_add_collaborator}
\alias{acl_add_viewer}
\alias{acl_remove_user}
\alias{acl_remove_collaborator}
\alias{acl_remove_viewer}
\alias{acl_add_self}
\alias{acl_remove_self}
\title{ACL Add User}
\usage{
acl_add_user(content, user_guid, role)

acl_add_collaborator(content, user_guid)

acl_add_viewer(content, user_guid)

acl_remove_user(content, user_guid)

acl_remove_collaborator(content, user_guid)

acl_remove_viewer(content, user_guid)

acl_add_self(content)

acl_remove_self(content)
}
\arguments{
\item{content}{The R6 Content object}

\item{user_guid}{The user's GUID}

\item{role}{One of "owner" or "viewer"}
}
\value{
The R6 content object (for piping)
}
\description{
Add a user_guid to the content as an owner or viewer
}
\keyword{internal}
