% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrations.R
\name{as_integration}
\alias{as_integration}
\title{Convert objects to integration class}
\usage{
as_integration(x, ...)
}
\arguments{
\item{x}{An object to convert to an integration}

\item{...}{Additional arguments passed to methods}
}
\value{
An integration object
}
\description{
Convert objects to integration class
}
