% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content.R
\name{as_tibble.connect_content_list}
\alias{as_tibble.connect_content_list}
\title{Convert integration list to a tibble}
\usage{
\method{as_tibble}{connect_content_list}(x, ...)
}
\arguments{
\item{x}{A \code{connect_content_list} object.}

\item{...}{Unused.}
}
\value{
A tibble with one row per content item.
}
\description{
Converts a list returned by \code{\link[=search_content]{search_content()}} to a tibble.
}
