% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrations.R
\name{as_tibble.connect_integration_list}
\alias{as_tibble.connect_integration_list}
\title{Convert integration list to a tibble}
\usage{
\method{as_tibble}{connect_integration_list}(x, ...)
}
\arguments{
\item{x}{A \code{connect_integration_list} object.}

\item{...}{Unused.}
}
\value{
A tibble with one row per integration.
}
\description{
Converts a list returned by \code{\link[=get_integrations]{get_integrations()}} to a tibble.
}
