% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{as_tibble.connect_list_hits}
\alias{as_tibble.connect_list_hits}
\title{Convert usage data to a tibble}
\usage{
\method{as_tibble}{connect_list_hits}(x, ...)
}
\arguments{
\item{x}{A \code{connect_list_hits} object.}

\item{...}{Passed to \code{\link[=as.data.frame]{as.data.frame()}}.}
}
\value{
A tibble with one row per usage record.
}
\description{
Converts an object returned by \code{\link[=get_usage]{get_usage()}} to a tibble via
\code{\link[=as.data.frame.connect_list_hits]{as.data.frame.connect_list_hits()}}.
}
