% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browse.R
\name{browse_solo}
\alias{browse_solo}
\alias{browse_dashboard}
\alias{browse_api_docs}
\title{Browse}
\usage{
browse_solo(content)

browse_dashboard(content)

browse_api_docs(connect)
}
\arguments{
\item{content}{A R6 Content object}

\item{connect}{A R6 Connect object}
}
\value{
The url that is opened in the browser
}
\description{
Browse to different locations on Connect via 
utils::browseURL
}
