% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy.R
\name{bundle_path}
\alias{bundle_path}
\title{Define a bundle from a path (a path directly to a tar.gz file)}
\usage{
bundle_path(path)
}
\arguments{
\item{path}{The path to a .tar.gz file}
}
\value{
Bundle A bundle object
}
\description{
Define a bundle from a path (a path directly to a tar.gz file)
}
\examples{
\dontshow{if (identical(Sys.getenv("IN_PKGDOWN"), "true")) withAutoprint(\{ # examplesIf}

bundle_path(system.file("tests/testthat/examples/static.tar.gz", package = "connectapi"))
\dontshow{\}) # examplesIf}
}
\seealso{
Other deployment functions:
\code{\link[=bundle_dir]{bundle_dir()}},
\code{\link[=bundle_static]{bundle_static()}},
\code{\link[=deploy]{deploy()}},
\code{\link[=download_bundle]{download_bundle()}},
\code{\link[=poll_task]{poll_task()}}
}
\concept{deployment functions}
