% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content.R
\name{create_random_name}
\alias{create_random_name}
\title{Create Random Name}
\usage{
create_random_name(length = 25)
}
\arguments{
\item{length}{Optional. The length of the random name. Defaults to 25}
}
\value{
The randomly generated name
}
\description{
Creates a random name from the LETTERS dataset
}
\seealso{
connectapi::verify_content_name

Other content functions:
\code{\link[=content_delete]{content_delete()}},
\code{\link[=content_item]{content_item()}},
\code{\link[=content_title]{content_title()}},
\code{\link[=content_update]{content_update()}},
\code{\link[=dashboard_url]{dashboard_url()}},
\code{\link[=delete_thumbnail]{delete_thumbnail()}},
\code{\link[=delete_vanity_url]{delete_vanity_url()}},
\code{\link[=deploy_repo]{deploy_repo()}},
\code{\link[=get_associations]{get_associations()}},
\code{\link[=get_bundles]{get_bundles()}},
\code{\link[=get_environment]{get_environment()}},
\code{\link[=get_jobs]{get_jobs()}},
\code{\link[=get_log]{get_log()}},
\code{\link[=get_thumbnail]{get_thumbnail()}},
\code{\link[=get_vanity_url]{get_vanity_url()}},
\code{\link{git}},
\code{\link[=has_thumbnail]{has_thumbnail()}},
\code{\link[=lock_content]{lock_content()}},
\code{\link{permissions}},
\code{\link[=search_content]{search_content()}},
\code{\link[=set_integrations]{set_integrations()}},
\code{\link[=set_run_as]{set_run_as()}},
\code{\link[=set_thumbnail]{set_thumbnail()}},
\code{\link[=set_vanity_url]{set_vanity_url()}},
\code{\link[=swap_vanity_urls]{swap_vanity_urls()}},
\code{\link[=terminate_jobs]{terminate_jobs()}},
\code{\link[=verify_content_name]{verify_content_name()}}
}
\concept{content functions}
