% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy.R
\name{dashboard_url_chr}
\alias{dashboard_url_chr}
\title{Build a Dashboard URL from Character Vectors}
\usage{
dashboard_url_chr(connect_url, content_guid, pane = "")
}
\arguments{
\item{connect_url}{character The base URL of the Connect server}

\item{content_guid}{character The guid for the content item in question}

\item{pane}{character The pane in the dashboard to link to}
}
\value{
character The dashboard URL for the content provided
}
\description{
Returns the URL for the content dashboard (opened to the selected pane).
NOTE: this takes a character object for performance optimization.
}
\seealso{
Other content functions: 
\code{\link{acl_add_user}()},
\code{\link{content_item}()},
\code{\link{content_title}()},
\code{\link{create_random_name}()},
\code{\link{dashboard_url}()},
\code{\link{deploy_repo}()},
\code{\link{get_acl_user}()},
\code{\link{get_bundles}()},
\code{\link{get_environment}()},
\code{\link{get_image}()},
\code{\link{get_jobs}()},
\code{\link{get_vanity_url}()},
\code{\link{git}},
\code{\link{set_image_path}()},
\code{\link{set_run_as}()},
\code{\link{set_vanity_url}()},
\code{\link{swap_vanity_url}()},
\code{\link{verify_content_name}()}
}
\concept{content functions}
