% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execution-envs.R
\name{ExecutionEnv}
\alias{ExecutionEnv}
\title{Execution Environment}
\description{
An R6 class representing an execution environment.
}
\seealso{
Other R6 classes: 
\code{\link{Bundle}},
\code{\link{Content}},
\code{\link{ContentTask}},
\code{\link{Environment}},
\code{\link{PositConnect}},
\code{\link{Task}},
\code{\link{Vanity}},
\code{\link{Variant}},
\code{\link{VariantSchedule}},
\code{\link{VariantTask}}
}
\concept{R6 classes}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ExecutionEnv-new}{\code{ExecutionEnv$new()}}
\item \href{#method-ExecutionEnv-print}{\code{ExecutionEnv$print()}}
\item \href{#method-ExecutionEnv-clone}{\code{ExecutionEnv$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExecutionEnv-new"></a>}}
\if{latex}{\out{\hypertarget{method-ExecutionEnv-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExecutionEnv$new(connect, data)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExecutionEnv-print"></a>}}
\if{latex}{\out{\hypertarget{method-ExecutionEnv-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExecutionEnv$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExecutionEnv-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ExecutionEnv-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExecutionEnv$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
