% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content.R
\name{get_acl}
\alias{get_acl}
\title{Get ACL Details}
\usage{
get_acl(content)
}
\arguments{
\item{content}{\link{Content} An R6 Content item}
}
\value{
A list of users who have access to the content
}
\description{
\lifecycle{experimental} Retrieve the Access Controls
associated with a given piece of content.
}
\seealso{
Other content functions: \code{\link{content_item}},
  \code{\link{dashboard_url_chr}},
  \code{\link{dashboard_url}},
  \code{\link{get_vanity_url}},
  \code{\link{set_image_path}},
  \code{\link{set_vanity_url}}
}
\concept{content functions}
