% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_aws_content_credentials}
\alias{get_aws_content_credentials}
\title{Perform an OAuth credential exchange to obtain AWS credentials for your content.}
\usage{
get_aws_content_credentials(connect, content_session_token = NULL)
}
\arguments{
\item{connect}{A Connect R6 object.}
}
\value{
The AWS credentials.
}
\description{
Perform an OAuth credential exchange to obtain AWS credentials for your content.
}
\details{
Please see https://docs.posit.co/connect/user/oauth-integrations/#obtaining-service-account-aws-credentials
for more information.
}
\examples{
\dontrun{
library(connectapi)
library(plumber)
library(paws)
client <- connect()

#* @get /do
function(req) {
  user_session_token <- req$HTTP_POSIT_CONNECT_USER_SESSION_TOKEN
  aws_credentials <- get_aws_content_credentials(client)

  # Create S3 client with AWS credentials from Connect 
  svc <- paws::s3(
    credentials = list(
      creds = list(
        access_key_id = aws_credentials$accessKeyId,
        secret_access_key = aws_credentials$secretAccessKey, 
        session_token = aws_credentials$sessionToken
      )
    )
  )

  # Get object from S3
  obj <- svc$get_object(
    Bucket = "my-bucket",
    Key = "my-data.csv"
  )

  "done"
}
}

}
