% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrations.R
\name{get_content_list}
\alias{get_content_list}
\title{Get content items using an OAuth integration}
\usage{
get_content_list(x)
}
\arguments{
\item{x}{A \code{connect_integration} object (as returned by \code{\link[=get_integrations]{get_integrations()}},
\code{\link[=get_integration]{get_integration()}}, \code{\link[=create_integration]{create_integration()}}, or \code{\link[=update_integration]{update_integration()}}).}
}
\value{
A list of \code{Content} R6 objects representing all pieces of content
that use the specified OAuth integration. See \code{\link[=content_item]{content_item()}} for details
on the object.
}
\description{
Retrieves a list of content items that are associated with a specific OAuth
integration.

You must have administrator privileges to use this function.
}
\examples{
\dontrun{
client <- connect()

# Get an integration
integration <- get_integration(client, "12345678-90ab-cdef-1234-567890abcdef")

# Find all content using this integration
content_using_integration <- get_content_list(integration)
}

}
\seealso{
\code{\link[=get_integrations]{get_integrations()}}, \code{\link[=get_integration]{get_integration()}}, \code{\link[=get_associations]{get_associations()}},
\code{\link[=content_item]{content_item()}}

Other oauth integration functions: 
\code{\link{create_integration}()},
\code{\link{delete_integration}()},
\code{\link{get_associations}()},
\code{\link{get_integration}()},
\code{\link{get_integrations}()},
\code{\link{set_integrations}()},
\code{\link{update_integration}()}
}
\concept{oauth integration functions}
