% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy.R
\name{get_image}
\alias{get_image}
\title{Get the Image}
\usage{
get_image(content, path)
}
\arguments{
\item{content}{A content object}

\item{path}{The path to the image on disk}
}
\description{
Get the Image
}
\seealso{
Other deployment functions: \code{\link{bundle_dir}},
  \code{\link{bundle_path}}, \code{\link{deploy}},
  \code{\link{download_bundle}}, \code{\link{poll_task}}
}
\concept{deployment functions}
