% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrations.R
\name{get_oauth_associations}
\alias{get_oauth_associations}
\title{Get OAuth integration associations for a piece of content}
\usage{
get_oauth_associations(content)
}
\arguments{
\item{content}{A \code{Content} R6 object representing the content item.}
}
\value{
A list of OAuth integration associations. Each association includes details such as:
\itemize{
\item \code{app_guid}: The content item's GUID (deprecated, use \code{content_guid} instead).
\item \code{content_guid}: The content item's GUID.
\item \code{oauth_integration_guid}: The GUID of the OAuth integration.
\item \code{oauth_integration_name}: The name of the OAuth integration.
\item \code{oauth_integration_description}: A description of the OAuth integration.
\item \code{oauth_integration_template}: The template used for this OAuth integration.
\item \code{oauth_integration_auth_type}: The authentication type (e.g., "Viewer" or "Service Account").
\item \code{created_time}: The timestamp when the association was created.
}
}
\description{
Retrieves a list of all OAuth integration associations for a content item.
These associations represent the OAuth integrations that have been set up for
the content to authenticate with external services.
}
\examples{
\dontrun{
client <- connect()
content <- content_item(client, "12345678-90ab-cdef-1234-567890abcdef")
associations <- get_oauth_associations(content)
}

}
\seealso{
\code{\link[=set_integrations]{set_integrations()}}, \code{\link[=content_get_integrations]{content_get_integrations()}}, \code{\link[=get_integrations]{get_integrations()}}

Other oauth integration functions: 
\code{\link{get_integration}()},
\code{\link{set_integrations}()}

Other content functions: 
\code{\link{content_delete}()},
\code{\link{content_item}()},
\code{\link{content_title}()},
\code{\link{content_update}()},
\code{\link{create_random_name}()},
\code{\link{dashboard_url}()},
\code{\link{delete_thumbnail}()},
\code{\link{delete_vanity_url}()},
\code{\link{deploy_repo}()},
\code{\link{get_bundles}()},
\code{\link{get_environment}()},
\code{\link{get_image}()},
\code{\link{get_job}()},
\code{\link{get_jobs}()},
\code{\link{get_log}()},
\code{\link{get_thumbnail}()},
\code{\link{get_vanity_url}()},
\code{\link{git}},
\code{\link{has_thumbnail}()},
\code{\link{permissions}},
\code{\link{set_image_path}()},
\code{\link{set_integrations}()},
\code{\link{set_run_as}()},
\code{\link{set_thumbnail}()},
\code{\link{set_vanity_url}()},
\code{\link{swap_vanity_url}()},
\code{\link{swap_vanity_urls}()},
\code{\link{terminate_jobs}()},
\code{\link{verify_content_name}()}
}
\concept{content functions}
\concept{oauth integration functions}
