% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_procs}
\alias{get_procs}
\title{Get Real-Time Process Data}
\usage{
get_procs(src)
}
\arguments{
\item{src}{The source object}
}
\value{
A tibble with the following columns:
\itemize{
\item \code{pid}: The PID of the current process
\item \code{appId}: The application ID
\item \code{appGuid}: The application GUID
\item \code{appName}: The application name
\item \code{appUrl}: The application URL
\item \code{appRunAs}: The application RunAs user
\item \code{type}: The type of process
\item \code{cpuCurrent}: The current CPU usage
\item \code{cpuTotal}: The total CPU usage
\item \code{ram}: The current RAM usage
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
This returns real-time process data from the Posit Connect API. It requires
administrator privileges to use. NOTE that this only returns data for the
server that responds to the request (i.e. in a Highly Available cluster)
}
