% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags.R
\name{get_tag_info}
\alias{get_tag_info}
\title{Get a tibble with more information for each tag}
\usage{
get_tag_info(src)
}
\arguments{
\item{src}{The source object}
}
\description{
Get a tibble with more information for each tag
}
