% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags.R
\name{get_tag_tree}
\alias{get_tag_tree}
\title{Print a visualization of the tag tree}
\usage{
get_tag_tree(src, top_tag = "RStudio Connect Tags")
}
\arguments{
\item{src}{The source object}

\item{top_tag}{The value to be printed at the top of the
tag tree. This value can not have any "/" characters in it.}
}
\description{
Print a visualization of the tag tree
}
