% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags.R
\name{get_tags}
\alias{get_tags}
\title{Get a simple data.frame listing all Tags on server}
\usage{
get_tags(src, use_cache = FALSE)
}
\arguments{
\item{src}{The source object}

\item{use_cache}{use the tag list previously queried
from the connect server. src$tags object is null then
this parameter will be ignored}
}
\description{
Get a simple data.frame listing all Tags on server
}
