% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schedule.R
\name{get_timezones}
\alias{get_timezones}
\title{Get TimeZones}
\usage{
get_timezones(connect)
}
\arguments{
\item{connect}{An R6 Connect object}
}
\value{
A TimeZone vector to be used for setting time zones
}
\description{
Get the available timezones from the server.
}
\seealso{
Other schedule functions:
\code{\link[=get_variant_schedule]{get_variant_schedule()}},
\code{\link[=set_schedule]{set_schedule()}}
}
\concept{schedule functions}
