% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{get_user_guid}
\alias{get_user_guid}
\title{Extract User GUID}
\usage{
get_user_guid(user)
}
\arguments{
\item{user}{Either a character string containing a user GUID or a
\code{connect_user} object (as returned by \code{Connect$user()} or \code{Connect$users()})}
}
\value{
A character string containing the user GUID
}
\description{
Helper function to extract a user GUID from either a character string or a
\code{connect_user} object.
}
\keyword{internal}
