% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_vanity_urls}
\alias{get_vanity_urls}
\title{Get all vanity URLs}
\usage{
get_vanity_urls(client)
}
\arguments{
\item{client}{A \code{Connect} object.}
}
\value{
A tibble with columns for \code{content_guid}, \code{path}, and
\code{created_time}.
}
\description{
Get a table of all vanity URLs on the server. Requires administrator
privileges.
}
\examples{
\dontrun{
library(connectapi)
client <- connect()
get_vanity_urls(client)
}

}
