% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schedule.R
\name{get_variant_schedule}
\alias{get_variant_schedule}
\title{Get a Variant Schedule}
\usage{
get_variant_schedule(variant)
}
\arguments{
\item{variant}{A Variant object, as returned by \code{get_variant()} or \code{get_variant_default()}}
}
\value{
A VariantSchedule object
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}} Gets the schedule associated with a Variant.
}
\seealso{
Other schedule functions:
\code{\link[=get_timezones]{get_timezones()}},
\code{\link[=set_schedule]{set_schedule()}}
}
\concept{schedule functions}
