% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content.R
\name{get_jobs}
\alias{get_jobs}
\alias{get_job}
\title{Get Jobs}
\usage{
get_jobs(content)

get_job(content, key)
}
\arguments{
\item{content}{A Content object, as returned by \code{content_item()}}

\item{key}{The key for a job}
}
\description{
\lifecycle{experimental} Retrieve details about jobs associated with a \code{content_item}.
"Jobs" in RStudio Connect are content executions
}
\seealso{
Other content functions: 
\code{\link{acl_add_user}()},
\code{\link{content_item}()},
\code{\link{content_title}()},
\code{\link{create_random_name}()},
\code{\link{dashboard_url_chr}()},
\code{\link{dashboard_url}()},
\code{\link{deploy_repo}()},
\code{\link{get_acl_user}()},
\code{\link{get_bundles}()},
\code{\link{get_environment}()},
\code{\link{get_image}()},
\code{\link{get_vanity_url}()},
\code{\link{git}},
\code{\link{set_image_path}()},
\code{\link{set_run_as}()},
\code{\link{set_vanity_url}()},
\code{\link{swap_vanity_url}()},
\code{\link{verify_content_name}()}
}
\concept{content functions}
