% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy.R
\name{poll_task}
\alias{poll_task}
\title{Poll Task}
\usage{
poll_task(task, wait = 1, callback = message)
}
\arguments{
\item{task}{A Task object}

\item{wait}{The interval to wait between polling}

\item{callback}{A function to be called for each message received. Set to NULL for no callback}
}
\value{
Task The Task object that was input
}
\description{
Polls a task, waiting for information about a deployment. If the task has
results, the output will be a modified "Task" object with \code{task$get_data()}
available to retrieve the results.
}
\details{
For a simple way to silence messages, set \code{callback = NULL}
}
\seealso{
Other deployment functions:
\code{\link[=bundle_dir]{bundle_dir()}},
\code{\link[=bundle_path]{bundle_path()}},
\code{\link[=bundle_static]{bundle_static()}},
\code{\link[=deploy]{deploy()}},
\code{\link[=download_bundle]{download_bundle()}}
}
\concept{deployment functions}
