% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/promote.R
\name{promote}
\alias{promote}
\title{Promote content from one Connect server to another}
\usage{
promote(from, to, to_key, from_key, app_name)
}
\arguments{
\item{from}{The url for the server containing the content (the originating
server)}

\item{to}{The url for the server where the content will be deployed (the
destination server)}

\item{to_key}{An API key on the destination "to" server. If the destination
content is going to be updated, the API key must belong to a user with
collaborator access on the content that will be updated. If the destination
content is to be created new, the API key must belong to a user with
publisher priviliges.}

\item{from_key}{An API key on the originating "from" server. The API key must
belong to a user with collaborator access to the content to be promoted.}

\item{app_name}{The name of the content on the originating "from" server.
If content with the same name is found on the destination server,
the content will be updated. If no content on the destination server
has a matching name, a new endpoint will be created.}
}
\value{
The URL for the content on the destination "to" server
}
\description{
Promote content from one Connect server to another
}
