% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\docType{data}
\name{RStudioConnect}
\alias{RStudioConnect}
\alias{Connect}
\title{Class representing a Connect API client}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Connect
}
\description{
Class representing a Connect API client
}
\section{Usage}{

\preformatted{
client <- Connect$new(host = 'connect.example.com',
  apiKey = 'mysecretkey')
client$get_apps()
client$get_tags()
}
}

\section{Details}{


This class allows a user to interact with a Connect server via the Connect
API. Authentication is done by providing an API key.
}

\keyword{datasets}
