% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy.R
\name{set_content_image}
\alias{set_content_image}
\alias{set_image_webshot}
\alias{set_image_path}
\alias{set_image_url}
\title{Set the Content Image}
\usage{
set_content_image(content, path)

set_image_webshot(content, ...)

set_image_webshot(content, ...)

set_image_path(content, path)

set_image_url(content, path)
}
\arguments{
\item{content}{A content object}

\item{path}{A file path or URL to an image}
}
\description{
\lifecycle{experimental}
}
\details{
Set the Content Image using a variety of methods.
}
\seealso{
Other content functions: 
\code{\link{content_delete}()},
\code{\link{content_item}()},
\code{\link{content_title}()},
\code{\link{content_update}()},
\code{\link{create_random_name}()},
\code{\link{dashboard_url}()},
\code{\link{dashboard_url_chr}()},
\code{\link{delete_vanity_url}()},
\code{\link{deploy_repo}()},
\code{\link{get_bundles}()},
\code{\link{get_environment}()},
\code{\link{get_image}()},
\code{\link{get_jobs}()},
\code{\link{get_vanity_url}()},
\code{\link{git}},
\code{\link{permissions}},
\code{\link{set_run_as}()},
\code{\link{set_vanity_url}()},
\code{\link{swap_vanity_url}()},
\code{\link{verify_content_name}()}
}
\concept{content functions}
