% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag_page.R
\name{tag_page}
\alias{tag_page}
\title{Generate a landing page for a specific RStudio Connect tag}
\usage{
tag_page(server, server_key, tag, description = NULL)
}
\arguments{
\item{server}{Url for the Connect server}

\item{server_key}{An API key for the Connect server. To get screen shots of
each app, the owner of the API key must be a viewer for each piece of
content.}

\item{tag}{The name of the targeted tag}

\item{description}{A description of the tag, placed at the top of the landing
page}
}
\value{
A list with LANDING_PAGE which is the path to the html file and APPS
  which is a list containing information on the apps and screenshots, useful
  if you'd rather use your own template
}
\description{
Generate a landing page for a specific RStudio Connect tag
}
