% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content.R
\name{terminate_jobs}
\alias{terminate_jobs}
\title{Terminate Jobs}
\usage{
terminate_jobs(content, keys = NULL)
}
\arguments{
\item{content}{A Content object, as returned by \code{content_item()}}

\item{keys}{Optional. One or more job keys, which can be obtained using
\code{get_jobs(content)}. If no keys are provided, will terminate all active
jobs for the provided content item.}
}
\value{
A data frame with the status of each termination request.
\itemize{
\item \code{app_id}: The content item's identifier.
\item \code{app_guid}: The content item's GUID.
\item \code{job_key}: The job key.
\item \code{job_id}: The job's identifier.
\item \code{result}: The result string returned by Connect.
\item \code{code}: An error code, \code{NA} if the request was successful.
\item \code{error}: An error message, \code{NA} if the result was successful.
}

Note that \code{app_id}, \code{app_guid}, \code{job_id}, and \code{result} are \code{NA} if the
request returns an error.
}
\description{
Register a job kill order for one or more jobs associated with a content
item. Requires Connect 2022.10.0 or newer.
}
\examples{
\dontrun{
client <- connect()
item <- content_item(client, "951bf3ad-82d0-4bca-bba8-9b27e35c49fa")
result <- terminate_jobs(item)
}

}
\seealso{
Other job functions:
\code{\link[=get_jobs]{get_jobs()}},
\code{\link[=get_log]{get_log()}}

Other content functions:
\code{\link[=content_delete]{content_delete()}},
\code{\link[=content_item]{content_item()}},
\code{\link[=content_title]{content_title()}},
\code{\link[=content_update]{content_update()}},
\code{\link[=create_random_name]{create_random_name()}},
\code{\link[=dashboard_url]{dashboard_url()}},
\code{\link[=delete_thumbnail]{delete_thumbnail()}},
\code{\link[=delete_vanity_url]{delete_vanity_url()}},
\code{\link[=deploy_repo]{deploy_repo()}},
\code{\link[=get_associations]{get_associations()}},
\code{\link[=get_bundles]{get_bundles()}},
\code{\link[=get_environment]{get_environment()}},
\code{\link[=get_jobs]{get_jobs()}},
\code{\link[=get_log]{get_log()}},
\code{\link[=get_thumbnail]{get_thumbnail()}},
\code{\link[=get_vanity_url]{get_vanity_url()}},
\code{\link{git}},
\code{\link[=has_thumbnail]{has_thumbnail()}},
\code{\link[=lock_content]{lock_content()}},
\code{\link{permissions}},
\code{\link[=search_content]{search_content()}},
\code{\link[=set_integrations]{set_integrations()}},
\code{\link[=set_run_as]{set_run_as()}},
\code{\link[=set_thumbnail]{set_thumbnail()}},
\code{\link[=set_vanity_url]{set_vanity_url()}},
\code{\link[=swap_vanity_urls]{swap_vanity_urls()}},
\code{\link[=verify_content_name]{verify_content_name()}}
}
\concept{content functions}
\concept{job functions}
