% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{user_guid_from_username}
\alias{user_guid_from_username}
\title{User}
\usage{
user_guid_from_username(client, username)
}
\arguments{
\item{client}{A Connect R6 object}

\item{username}{The user's username}
}
\description{
Get user details
}
\details{
\code{user_guid_from_username()} is a helper to retrieve a user GUID, given the
user's username. It is useful in Shiny applications for using \code{session$user}
}
