% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant.R
\name{get_variants}
\alias{get_variants}
\alias{get_variant}
\alias{get_variant_default}
\title{Get Variant}
\usage{
get_variants(content)

get_variant(content, key)

get_variant_default(content)
}
\arguments{
\item{content}{An R6 Content object. Returned from \code{content_item()}}

\item{key}{The Variant key for a specific variant}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}} Work with variants
}
\details{
\itemize{
\item \code{get_variants()} returns a \code{tibble} with variant data for a \code{content_item}
\item \code{get_variant_default()} returns the default variant for a \code{content_item}
\item \code{get_variant()} returns a specific variant for a \code{content_item} (specified by \code{key})
}
}
\seealso{
Other variant functions:
\code{\link[=get_variant_renderings]{get_variant_renderings()}}
}
\concept{variant functions}
