% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant.R
\name{get_variant_renderings}
\alias{get_variant_renderings}
\alias{variant_render}
\title{Render a Variant}
\usage{
get_variant_renderings(variant)

variant_render(variant)
}
\arguments{
\item{variant}{An R6 Variant object. As returned by \code{get_variant()} or \code{get_variant_default()}}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}} Get details about renderings (i.e. render history)
or execute a variant on demand
}
\details{
\itemize{
\item \code{get_variant_renderings()} returns all renderings / content for a particular variant. Returns a \code{tibble}
\item \code{variant_render()} executes a variant on demand. Returns a \code{VariantTask} object
}
}
\seealso{
Other variant functions:
\code{\link[=get_variants]{get_variants()}}
}
\concept{variant functions}
