% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-ci.R
\name{wait_for_process}
\alias{wait_for_process}
\title{Wait for a Process to Complete}
\usage{
wait_for_process(proc)
}
\arguments{
\item{proc}{A processx process object}
}
\value{
A list with named stdout and stderr entries
}
\description{
It is important to poll output intermittently in case pipe buffers fill up.
Otherwise the process will be paused until the buffer is cleared.
}
\keyword{internal}
