% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrations.R
\name{as_integration}
\alias{as_integration}
\title{Convert objects to integration class}
\usage{
as_integration(x, client)
}
\arguments{
\item{x}{An object to convert to an integration.}

\item{client}{The Connect client object where the integration comes from.}
}
\value{
An integration object. The object has all the fields from the
integrations endpoint (see \code{\link[=get_integrations]{get_integrations()}}) and a Connect client as a
\code{client} attribute (\code{attr(x, "client")})
}
\description{
Convert objects to integration class
}
