% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrations.R
\name{as_tibble.connect_list_integrations}
\alias{as_tibble.connect_list_integrations}
\title{Convert integrations data to a tibble}
\usage{
\method{as_tibble}{connect_list_integrations}(x, ...)
}
\arguments{
\item{x}{A \code{connect_list_integrations} object.}

\item{...}{Unused.}
}
\value{
A tibble with one row per integration.
}
\description{
Converts a list returned by \code{\link[=get_integrations]{get_integrations()}} to a tibble.
}
