% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audits.R
\name{audit_access_open}
\alias{audit_access_open}
\title{Audit Access Controls}
\usage{
audit_access_open(apps, type = "all")
}
\arguments{
\item{apps}{App list, see \code{cache_apps}}

\item{type}{One of "all" or "logged_in". If "all", return a list of apps
whose access control is set to "Everyone". If "logged_in", return a list of
apps whose access control is set to "All logged in users"}
}
\description{
Audit Access Controls
}
\seealso{
Other audit functions: \code{\link{audit_r_versions}},
  \code{\link{audit_runas}},
  \code{\link{audit_vanity_urls}}, \code{\link{cache_apps}}
}
\concept{audit functions}
