% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audits.R
\name{audit_access_open}
\alias{audit_access_open}
\title{Audit Access Controls}
\usage{
audit_access_open(content, type = "all")
}
\arguments{
\item{content}{\code{data.frame} of content information, as from \code{\link[=get_content]{get_content()}}}

\item{type}{One of "all" or "logged_in". If "all", return a list of apps
whose access control is set to "Everyone". If "logged_in", return a list of
apps whose access control is set to "All logged in users"}
}
\description{
\lifecycle{experimental}
}
\seealso{
Other audit functions: 
\code{\link{audit_r_versions}()},
\code{\link{audit_runas}()},
\code{\link{vanity_is_available}()}
}
\concept{audit functions}
