% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audits.R
\name{audit_r_versions}
\alias{audit_r_versions}
\title{Audit R Versions}
\usage{
audit_r_versions(content)
}
\arguments{
\item{content}{\code{data.frame} of content information, as from \code{\link[=get_content]{get_content()}}}
}
\value{
A plot that shows the R version used by content over time and in
aggregate.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\seealso{
Other audit functions: 
\code{\link{audit_access_open}()},
\code{\link{audit_runas}()},
\code{\link{vanity_is_available}()}
}
\concept{audit functions}
