% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audits.R
\name{audit_r_versions}
\alias{audit_r_versions}
\title{Audit R Versions}
\usage{
audit_r_versions(apps)
}
\arguments{
\item{apps}{App list, see \code{cache_apps}}
}
\value{
A plot that shows the R version used by content over time and in
  aggregate.
}
\description{
Audit R Versions
}
