% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audits.R
\name{audit_runas}
\alias{audit_runas}
\title{Audit Run As Settings}
\usage{
audit_runas(content)
}
\arguments{
\item{content}{\code{data.frame} of content information, as from \code{\link[=get_content]{get_content()}}}
}
\value{
A data frame with the app name and the Run As user if the Run As user
is not the default
}
\description{
\lifecycle{experimental}
}
\seealso{
Other audit functions: 
\code{\link{audit_access_open}()},
\code{\link{audit_r_versions}()},
\code{\link{vanity_is_available}()}
}
\concept{audit functions}
