% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audits.R
\name{audit_runas}
\alias{audit_runas}
\title{Audit Run As Settings}
\usage{
audit_runas(apps)
}
\arguments{
\item{apps}{App list, see \code{cache_apps}}
}
\value{
A data frame with the app name and the Run As user if the Run As user
  is not the default
}
\description{
Audit Run As Settings
}
