% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy.R
\docType{data}
\name{Bundle}
\alias{Bundle}
\title{Bundle}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Bundle
}
\description{
An S6 class that represents a bundle
}
\seealso{
Other deploy: \code{\link{Task}}, \code{\link{bundle_dir}},
  \code{\link{bundle_path}}, \code{\link{deploy}}
}
\concept{deploy}
\keyword{datasets}
