% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy.R
\docType{data}
\name{Bundle}
\alias{Bundle}
\title{Bundle}
\format{An object of class \code{R6ClassGenerator} of length 25.}
\usage{
Bundle
}
\description{
An R6 class that represents a bundle
}
\seealso{
Other R6 classes: \code{\link{Content}},
  \code{\link{RStudioConnect}}, \code{\link{Task}},
  \code{\link{Vanity}}
}
\concept{R6 classes}
\keyword{datasets}
