% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy.R
\name{bundle_path}
\alias{bundle_path}
\title{Define a bundle from a path (a tar.gz file)}
\usage{
bundle_path(connect, path)
}
\arguments{
\item{connect}{A Connect object}

\item{path}{The path to a .tar.gz file}
}
\value{
Bundle A bundle object
}
\description{
Define a bundle from a path (a tar.gz file)
}
\seealso{
Other deploy: \code{\link{Bundle}}, \code{\link{Task}},
  \code{\link{bundle_dir}}, \code{\link{deploy}}
}
\concept{deploy}
