% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy.R
\name{bundle_path}
\alias{bundle_path}
\title{Define a bundle from a path (a tar.gz file)}
\usage{
bundle_path(path)
}
\arguments{
\item{path}{The path to a .tar.gz file}
}
\value{
Bundle A bundle object
}
\description{
Define a bundle from a path (a tar.gz file)
}
\seealso{
Other deployment functions: 
\code{\link{bundle_dir}()},
\code{\link{bundle_static}()},
\code{\link{deploy}()},
\code{\link{download_bundle}()},
\code{\link{poll_task}()}
}
\concept{deployment functions}
