% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy.R
\name{bundle_static}
\alias{bundle_static}
\title{Define a bundle from a static file (or files)}
\usage{
bundle_static(
  path,
  filename = fs::file_temp(pattern = "bundle", ext = ".tar.gz")
)
}
\arguments{
\item{path}{The path to a file (or files) that will be used for the static bundle}
}
\value{
Bundle A bundle object
}
\description{
Define a bundle from a static file (or files)
}
\seealso{
Other deployment functions: 
\code{\link{bundle_dir}()},
\code{\link{bundle_path}()},
\code{\link{deploy}()},
\code{\link{download_bundle}()},
\code{\link{get_image}()},
\code{\link{poll_task}()}
}
\concept{deployment functions}
