% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audits.R
\name{cache_apps}
\alias{cache_apps}
\title{Get information on all apps for a server}
\usage{
cache_apps(connect)
}
\arguments{
\item{connect}{A Connect object}
}
\value{
List with application data, to be used by audit functions
}
\description{
Get information on all apps for a server
}
