% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audits.R
\name{cache_apps}
\alias{cache_apps}
\title{Get information on all apps for a server}
\usage{
cache_apps(server, server_key)
}
\arguments{
\item{server}{URL for Connect server}

\item{server_key}{Admin API Key for Connect server}
}
\value{
List with application data, to be used by audit functions
}
\description{
Get information on all apps for a server
}
