% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{connect}
\alias{connect}
\title{Create a connection to RStudio Connect}
\usage{
connect(host = Sys.getenv(paste0(prefix, "_SERVER"), NA_character_),
  api_key = Sys.getenv(paste0(prefix, "_API_KEY"), NA_character_),
  prefix = "RSTUDIO_CONNECT")
}
\arguments{
\item{host}{The URL for accessing RStudio Connect. Defaults to environment
variable RSTUDIO_CONNECT_SERVER}

\item{api_key}{The API Key to authenticate to RStudio Connect with. Defaults
to environment variable RSTUDIO_CONNECT_API_KEY}
}
\value{
An RStudio Connect R6 object that can be passed along to methods
}
\description{
Creates a connection to RStudio Connect using the hostname and an api key.
Validates the connection and checks that the version of the server is
compatible with the current version of the package.
}
