% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{Connect}
\alias{Connect}
\title{Class representing a Connect API client}
\description{
Class representing a Connect API client
}
\section{Usage}{

\preformatted{
client <- Connect$new(host = 'connect.example.com',
  apiKey = 'mysecretkey')
client$get_apps()
client$get_tags()
}
}

\section{Details}{


This class allows a user to interact with a Connect server via the Connect
API. Authentication is done by providing an API key.
}

