% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectapi-package.R, R/connectapi.R
\docType{package}
\name{connectapi-package}
\alias{connectapi-package}
\title{connectapi: Utilities for Interacting with the 'Posit Connect' Server API}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides a helpful 'R6' class and methods for interacting with the 'Posit Connect' Server API along with some meaningful utility functions for regular tasks. API documentation varies by 'Posit Connect' installation and version, but the latest documentation is also hosted publicly at \url{https://docs.posit.co/connect/api/}.

\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides a helpful 'R6' class and methods for interacting with the 'Posit Connect' Server API along with some meaningful utility functions for regular tasks. API documentation varies by 'Posit Connect' installation and version, but the latest documentation is also hosted publicly at \url{https://docs.posit.co/connect/api/}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://posit-dev.github.io/connectapi/}
  \item \url{https://github.com/posit-dev/connectapi}
  \item Report bugs at \url{https://github.com/posit-dev/connectapi/issues}
}


Useful links:
\itemize{
  \item \url{https://posit-dev.github.io/connectapi/}
  \item \url{https://github.com/posit-dev/connectapi}
  \item Report bugs at \url{https://github.com/posit-dev/connectapi/issues}
}

}
\author{
\strong{Maintainer}: Kara Woo \email{kara.woo@posit.co}

Authors:
\itemize{
  \item Toph Allen \email{toph@posit.co}
  \item Neal Richardson
  \item Sean Lopp
  \item Cole Arendt
}

Other contributors:
\itemize{
  \item Posit, PBC [copyright holder, funder]
}

}
\keyword{internal}
