% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectapi.R
\docType{package}
\name{connectapi-package}
\alias{connectapi}
\alias{connectapi-package}
\title{connectapi: Utilities for Interacting with the 'RStudio Connect' Server API}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides a helpful 'R6' class and methods for interacting with the 'RStudio Connect' Server API along with some meaningful utility functions for regular tasks. API documentation varies by 'RStudio Connect' installation and version, but the latest documentation is also hosted publicly at \url{https://docs.rstudio.com/connect/api/}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rstudio/connectapi}
}

}
\author{
\strong{Maintainer}: Cole Arendt \email{cole@rstudio.com}

Authors:
\itemize{
  \item Sean Lopp \email{sean@rstudio.com}
}

Other contributors:
\itemize{
  \item RStudio PBC [copyright holder, funder]
}

}
