% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectapi.R
\docType{package}
\name{connectapi-package}
\alias{connectapi}
\alias{connectapi-package}
\title{connectapi: Utilities for Interacting with the 'Posit Connect' Server API}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Provides a helpful 'R6' class and methods for interacting with the 'Posit Connect' Server API along with some meaningful utility functions for regular tasks. API documentation varies by 'Posit Connect' installation and version, but the latest documentation is also hosted publicly at <https://docs.posit.co/connect/api/>.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rstudio/connectapi}
}

}
\author{
\strong{Maintainer}: Cole Arendt \email{cole@posit.co}

Authors:
\itemize{
  \item Sean Lopp
}

Other contributors:
\itemize{
  \item Posit, PBC [copyright holder, funder]
}

}
