% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectapi.R
\docType{package}
\name{connectapi-package}
\alias{connectapi}
\alias{connectapi-package}
\title{connectapi: Utilities for Interacting with the RStudio Connect Server
    API}
\description{
\if{html}{\figure{logo.png}{options: align='right'}}

Helpful R6 class for interacting with the RStudio
    Connect Server API and some example utility functions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rstudio/connectapi}
}

}
\author{
\strong{Maintainer}: Cole Arendt \email{cole@rstudio.com}

Authors:
\itemize{
  \item Sean Lopp \email{sean@rstudio.com}
}

Other contributors:
\itemize{
  \item RStudio [copyright holder, funder]
}

}
