% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content.R
\docType{data}
\name{Content}
\alias{Content}
\title{Content}
\format{An object of class \code{R6ClassGenerator} of length 25.}
\usage{
Content
}
\description{
An R6 class that represents content
}
\section{Fields}{

\describe{
\item{\code{connect}}{An R6 Connect object}

\item{\code{content}}{The content details from RStudio Connect}
}}

\seealso{
Other R6 classes: \code{\link{Bundle}},
  \code{\link{RStudioConnect}}, \code{\link{Task}},
  \code{\link{Vanity}}
}
\concept{R6 classes}
\keyword{datasets}
