% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy.R
\docType{data}
\name{Content}
\alias{Content}
\title{Content}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Content
}
\description{
An R6 class that represents content
}
\seealso{
Other deployR6: \code{\link{Bundle}}, \code{\link{Task}},
  \code{\link{Vanity}}
}
\concept{deployR6}
\keyword{datasets}
