% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content.R
\name{Content}
\alias{Content}
\title{Content}
\description{
An R6 class that represents content
}
\seealso{
Other R6 classes: 
\code{\link{Bundle}},
\code{\link{ContentTask}},
\code{\link{RStudioConnect}},
\code{\link{Task}},
\code{\link{Vanity}},
\code{\link{VariantSchedule}},
\code{\link{VariantTask}}
}
\concept{R6 classes}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{connect}}{An R6 Connect object}

\item{\code{content}}{The content details from RStudio Connect}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Content$new()}}
\item \href{#method-get_connect}{\code{Content$get_connect()}}
\item \href{#method-get_content}{\code{Content$get_content()}}
\item \href{#method-get_content_remote}{\code{Content$get_content_remote()}}
\item \href{#method-get_bundles}{\code{Content$get_bundles()}}
\item \href{#method-internal_content}{\code{Content$internal_content()}}
\item \href{#method-update}{\code{Content$update()}}
\item \href{#method-runas}{\code{Content$runas()}}
\item \href{#method-get_url}{\code{Content$get_url()}}
\item \href{#method-get_dashboard_url}{\code{Content$get_dashboard_url()}}
\item \href{#method-get_jobs}{\code{Content$get_jobs()}}
\item \href{#method-get_job}{\code{Content$get_job()}}
\item \href{#method-jobs}{\code{Content$jobs()}}
\item \href{#method-job}{\code{Content$job()}}
\item \href{#method-variants}{\code{Content$variants()}}
\item \href{#method-tag_set}{\code{Content$tag_set()}}
\item \href{#method-tag_delete}{\code{Content$tag_delete()}}
\item \href{#method-tags}{\code{Content$tags()}}
\item \href{#method-permissions_add}{\code{Content$permissions_add()}}
\item \href{#method-permissions_update}{\code{Content$permissions_update()}}
\item \href{#method-permissions_delete}{\code{Content$permissions_delete()}}
\item \href{#method-permissions}{\code{Content$permissions()}}
\item \href{#method-environment}{\code{Content$environment()}}
\item \href{#method-environment_set}{\code{Content$environment_set()}}
\item \href{#method-deploy}{\code{Content$deploy()}}
\item \href{#method-repo_enable}{\code{Content$repo_enable()}}
\item \href{#method-repo_set}{\code{Content$repo_set()}}
\item \href{#method-print}{\code{Content$print()}}
\item \href{#method-clone}{\code{Content$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Content$new(connect, content)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_connect"></a>}}
\if{latex}{\out{\hypertarget{method-get_connect}{}}}
\subsection{Method \code{get_connect()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Content$get_connect()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_content"></a>}}
\if{latex}{\out{\hypertarget{method-get_content}{}}}
\subsection{Method \code{get_content()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Content$get_content()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_content_remote"></a>}}
\if{latex}{\out{\hypertarget{method-get_content_remote}{}}}
\subsection{Method \code{get_content_remote()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Content$get_content_remote()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_bundles"></a>}}
\if{latex}{\out{\hypertarget{method-get_bundles}{}}}
\subsection{Method \code{get_bundles()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Content$get_bundles()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-internal_content"></a>}}
\if{latex}{\out{\hypertarget{method-internal_content}{}}}
\subsection{Method \code{internal_content()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Content$internal_content()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update"></a>}}
\if{latex}{\out{\hypertarget{method-update}{}}}
\subsection{Method \code{update()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Content$update(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-runas"></a>}}
\if{latex}{\out{\hypertarget{method-runas}{}}}
\subsection{Method \code{runas()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Content$runas(run_as, run_as_current_user = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_url"></a>}}
\if{latex}{\out{\hypertarget{method-get_url}{}}}
\subsection{Method \code{get_url()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Content$get_url()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_dashboard_url"></a>}}
\if{latex}{\out{\hypertarget{method-get_dashboard_url}{}}}
\subsection{Method \code{get_dashboard_url()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Content$get_dashboard_url(pane = "")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_jobs"></a>}}
\if{latex}{\out{\hypertarget{method-get_jobs}{}}}
\subsection{Method \code{get_jobs()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Content$get_jobs()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_job"></a>}}
\if{latex}{\out{\hypertarget{method-get_job}{}}}
\subsection{Method \code{get_job()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Content$get_job(key)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-jobs"></a>}}
\if{latex}{\out{\hypertarget{method-jobs}{}}}
\subsection{Method \code{jobs()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Content$jobs()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-job"></a>}}
\if{latex}{\out{\hypertarget{method-job}{}}}
\subsection{Method \code{job()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Content$job(key)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-variants"></a>}}
\if{latex}{\out{\hypertarget{method-variants}{}}}
\subsection{Method \code{variants()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Content$variants()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-tag_set"></a>}}
\if{latex}{\out{\hypertarget{method-tag_set}{}}}
\subsection{Method \code{tag_set()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Content$tag_set(tag_id)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-tag_delete"></a>}}
\if{latex}{\out{\hypertarget{method-tag_delete}{}}}
\subsection{Method \code{tag_delete()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Content$tag_delete(id)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-tags"></a>}}
\if{latex}{\out{\hypertarget{method-tags}{}}}
\subsection{Method \code{tags()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Content$tags()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-permissions_add"></a>}}
\if{latex}{\out{\hypertarget{method-permissions_add}{}}}
\subsection{Method \code{permissions_add()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Content$permissions_add(principal_guid, principal_type, role)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-permissions_update"></a>}}
\if{latex}{\out{\hypertarget{method-permissions_update}{}}}
\subsection{Method \code{permissions_update()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Content$permissions_update(id, principal_guid, principal_type, role)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-permissions_delete"></a>}}
\if{latex}{\out{\hypertarget{method-permissions_delete}{}}}
\subsection{Method \code{permissions_delete()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Content$permissions_delete(id)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-permissions"></a>}}
\if{latex}{\out{\hypertarget{method-permissions}{}}}
\subsection{Method \code{permissions()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Content$permissions(id = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-environment"></a>}}
\if{latex}{\out{\hypertarget{method-environment}{}}}
\subsection{Method \code{environment()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Content$environment()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-environment_set"></a>}}
\if{latex}{\out{\hypertarget{method-environment_set}{}}}
\subsection{Method \code{environment_set()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Content$environment_set(..., .version = 0)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-deploy"></a>}}
\if{latex}{\out{\hypertarget{method-deploy}{}}}
\subsection{Method \code{deploy()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Content$deploy()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-repo_enable"></a>}}
\if{latex}{\out{\hypertarget{method-repo_enable}{}}}
\subsection{Method \code{repo_enable()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Content$repo_enable(enabled = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-repo_set"></a>}}
\if{latex}{\out{\hypertarget{method-repo_set}{}}}
\subsection{Method \code{repo_set()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Content$repo_set(repository, branch, subdirectory)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Content$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Content$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
