% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content.R
\name{content_delete}
\alias{content_delete}
\title{Delete Content}
\usage{
content_delete(content, force = FALSE)
}
\arguments{
\item{content}{an R6 content item}

\item{force}{Optional. A boolean that determines whether we should prompt in interactive sessions}
}
\value{
The R6 Content item. The item is deleted, but information about it is cached locally
}
\description{
Delete a content item. WARNING: This action deletes all history, configuration,
logs, and resources about a content item. It \emph{cannot} be undone.
}
\seealso{
Other content functions: 
\code{\link{content_item}()},
\code{\link{content_title}()},
\code{\link{content_update}()},
\code{\link{create_random_name}()},
\code{\link{dashboard_url}()},
\code{\link{dashboard_url_chr}()},
\code{\link{delete_thumbnail}()},
\code{\link{delete_vanity_url}()},
\code{\link{deploy_repo}()},
\code{\link{get_bundles}()},
\code{\link{get_environment}()},
\code{\link{get_image}()},
\code{\link{get_job}()},
\code{\link{get_jobs}()},
\code{\link{get_thumbnail}()},
\code{\link{get_vanity_url}()},
\code{\link{git}},
\code{\link{has_thumbnail}()},
\code{\link{permissions}},
\code{\link{set_image_path}()},
\code{\link{set_run_as}()},
\code{\link{set_thumbnail}()},
\code{\link{set_vanity_url}()},
\code{\link{swap_vanity_url}()},
\code{\link{verify_content_name}()}
}
\concept{content functions}
