% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content.R
\name{content_item}
\alias{content_item}
\title{Get Content Item}
\usage{
content_item(connect, guid)
}
\arguments{
\item{connect}{A Connect object}

\item{guid}{The GUID for the content item to be retrieved}
}
\value{
A Content object for use with other content endpoints
}
\description{
Returns a single content item based on guid
}
\seealso{
Other content functions: 
\code{\link{acl_add_user}()},
\code{\link{content_title}()},
\code{\link{dashboard_url_chr}()},
\code{\link{dashboard_url}()},
\code{\link{get_acl_user}()},
\code{\link{get_environment}()},
\code{\link{get_image}()},
\code{\link{get_jobs}()},
\code{\link{get_vanity_url}()},
\code{\link{set_image_path}()},
\code{\link{set_run_as}()},
\code{\link{set_vanity_url}()},
\code{\link{swap_vanity_url}()}
}
\concept{content functions}
