% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy.R
\name{content_item}
\alias{content_item}
\title{Get Content Item}
\usage{
content_item(connect, guid)
}
\arguments{
\item{connect}{A Connect object}

\item{guid}{The GUID for the content item to be retrieved}
}
\value{
A Content object for use with other content endpoints
}
\description{
Returns a single content item based on guid
}
\seealso{
Other content functions: \code{\link{get_vanity_url}},
  \code{\link{set_image_path}},
  \code{\link{set_vanity_url}}
}
\concept{content functions}
