% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content.R
\name{content_item}
\alias{content_item}
\title{Get Content Item}
\usage{
content_item(connect, guid)
}
\arguments{
\item{connect}{A Connect object}

\item{guid}{The GUID for the content item to be retrieved}
}
\value{
A Content object for use with other content endpoints
}
\description{
Returns a single content item based on guid
}
\examples{
\dontrun{
connect() \%>\%
  content_item("some-guid") \%>\%
  content_update_access_type("all")
}
}
\seealso{
Other content functions: 
\code{\link{content_delete}()},
\code{\link{content_title}()},
\code{\link{content_update}()},
\code{\link{create_random_name}()},
\code{\link{dashboard_url}()},
\code{\link{dashboard_url_chr}()},
\code{\link{delete_thumbnail}()},
\code{\link{delete_vanity_url}()},
\code{\link{deploy_repo}()},
\code{\link{get_bundles}()},
\code{\link{get_environment}()},
\code{\link{get_jobs}()},
\code{\link{get_thumbnail}()},
\code{\link{get_vanity_url}()},
\code{\link{git}},
\code{\link{has_thumbnail}()},
\code{\link{permissions}},
\code{\link{set_run_as}()},
\code{\link{set_thumbnail}()},
\code{\link{set_vanity_url}()},
\code{\link{swap_vanity_url}()},
\code{\link{verify_content_name}()}
}
\concept{content functions}
