% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content.R
\name{content_title}
\alias{content_title}
\title{Get Content Title}
\usage{
content_title(connect, guid, default = "Unknown Content")
}
\arguments{
\item{connect}{A Connect object}

\item{guid}{The GUID for the content item to be retrieved}

\item{default}{The default value returned for missing or not visible content}
}
\value{
character. The title of the requested content
}
\description{
Return content title for a piece of content. If the content
is missing (deleted) or not visible, then returns the \code{default}
}
\seealso{
Other content functions: 
\code{\link{content_delete}()},
\code{\link{content_item}()},
\code{\link{content_update}()},
\code{\link{create_random_name}()},
\code{\link{dashboard_url}()},
\code{\link{dashboard_url_chr}()},
\code{\link{delete_vanity_url}()},
\code{\link{deploy_repo}()},
\code{\link{get_bundles}()},
\code{\link{get_environment}()},
\code{\link{get_jobs}()},
\code{\link{get_thumbnail}()},
\code{\link{get_vanity_url}()},
\code{\link{git}},
\code{\link{permissions}},
\code{\link{set_run_as}()},
\code{\link{set_thumbnail}()},
\code{\link{set_vanity_url}()},
\code{\link{swap_vanity_url}()},
\code{\link{verify_content_name}()}
}
\concept{content functions}
