% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content.R
\name{content_update}
\alias{content_update}
\alias{content_access_type}
\title{Update Content}
\usage{
content_update(content, ...)

content_access_type(content, access_type = c("all", "logged_in", "acl"))
}
\arguments{
\item{content}{An R6 content item}

\item{...}{Settings up update that are passed along to RStudio Connect}

\item{access_type}{One of "all", "logged_in", or "acl"}
}
\value{
An R6 content item
}
\description{
Update settings for a content item. For a list of all settings, see the
\href{https://docs.rstudio.com/connect/api/#patch-/v1/content/{guid}}{latest documentation}
or the documentation for your server via \code{connectapi::browse_api_docs()}.
}
\details{
Popular selections are \code{content_update(access_type="all")},
\code{content_update(access_type="logged_in")} or
\code{content_update(access_type="acl")}, process settings, title, description,
etc.

\code{content_access_type()} is a helper to make this more discoverable
}
\seealso{
Other content functions: 
\code{\link{acl_add_user}()},
\code{\link{content_delete}()},
\code{\link{content_item}()},
\code{\link{content_title}()},
\code{\link{create_random_name}()},
\code{\link{dashboard_url_chr}()},
\code{\link{dashboard_url}()},
\code{\link{delete_vanity_url}()},
\code{\link{deploy_repo}()},
\code{\link{get_acl_user}()},
\code{\link{get_bundles}()},
\code{\link{get_environment}()},
\code{\link{get_image}()},
\code{\link{get_jobs}()},
\code{\link{get_vanity_url}()},
\code{\link{git}},
\code{\link{permissions}},
\code{\link{set_image_path}()},
\code{\link{set_run_as}()},
\code{\link{set_vanity_url}()},
\code{\link{swap_vanity_url}()},
\code{\link{verify_content_name}()}

Other content functions: 
\code{\link{acl_add_user}()},
\code{\link{content_delete}()},
\code{\link{content_item}()},
\code{\link{content_title}()},
\code{\link{create_random_name}()},
\code{\link{dashboard_url_chr}()},
\code{\link{dashboard_url}()},
\code{\link{delete_vanity_url}()},
\code{\link{deploy_repo}()},
\code{\link{get_acl_user}()},
\code{\link{get_bundles}()},
\code{\link{get_environment}()},
\code{\link{get_image}()},
\code{\link{get_jobs}()},
\code{\link{get_vanity_url}()},
\code{\link{git}},
\code{\link{permissions}},
\code{\link{set_image_path}()},
\code{\link{set_run_as}()},
\code{\link{set_vanity_url}()},
\code{\link{swap_vanity_url}()},
\code{\link{verify_content_name}()}
}
\concept{content functions}
