% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content.R
\name{create_random_name}
\alias{create_random_name}
\title{Create Random Name}
\usage{
create_random_name(length = 25)
}
\arguments{
\item{length}{Optional. The length of the random name. Defaults to 25}
}
\value{
The randomly generated name
}
\description{
Creates a random name from the LETTERS dataset
}
\seealso{
connectapi::verify_content_name

Other content functions: 
\code{\link{acl_add_user}()},
\code{\link{content_item}()},
\code{\link{content_title}()},
\code{\link{dashboard_url_chr}()},
\code{\link{dashboard_url}()},
\code{\link{get_acl_user}()},
\code{\link{get_bundles}()},
\code{\link{get_environment}()},
\code{\link{get_image}()},
\code{\link{get_jobs}()},
\code{\link{get_vanity_url}()},
\code{\link{set_image_path}()},
\code{\link{set_run_as}()},
\code{\link{set_vanity_url}()},
\code{\link{swap_vanity_url}()},
\code{\link{verify_content_name}()}
}
\concept{content functions}
