% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy.R
\name{dashboard_url}
\alias{dashboard_url}
\title{Build a Dashboard URL from a Content Item}
\usage{
dashboard_url(content, pane = "")
}
\arguments{
\item{content}{\link{R6:Content} A Content object}

\item{pane}{\link{character} The pane in the dashboard to link to}
}
\value{
\link{character} The dashboard URL for the content provided
}
\description{
Returns the URL for the content dashboard (opened to the selected pane).
}
\seealso{
Other content functions: \code{\link{content_item}},
  \code{\link{dashboard_url_chr}},
  \code{\link{get_vanity_url}},
  \code{\link{set_image_path}},
  \code{\link{set_vanity_url}}
}
\concept{content functions}
