% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy.R
\name{dashboard_url}
\alias{dashboard_url}
\title{Build a Dashboard URL from a Content Item}
\usage{
dashboard_url(content, pane = "")
}
\arguments{
\item{content}{\link{Content} A Content object}

\item{pane}{character The pane in the dashboard to link to}
}
\value{
character The dashboard URL for the content provided
}
\description{
Returns the URL for the content dashboard (opened to the selected pane).
}
\seealso{
Other content functions: 
\code{\link{content_delete}()},
\code{\link{content_item}()},
\code{\link{content_title}()},
\code{\link{content_update}()},
\code{\link{create_random_name}()},
\code{\link{dashboard_url_chr}()},
\code{\link{delete_thumbnail}()},
\code{\link{delete_vanity_url}()},
\code{\link{deploy_repo}()},
\code{\link{get_bundles}()},
\code{\link{get_environment}()},
\code{\link{get_jobs}()},
\code{\link{get_thumbnail}()},
\code{\link{get_vanity_url}()},
\code{\link{git}},
\code{\link{has_thumbnail}()},
\code{\link{permissions}},
\code{\link{set_run_as}()},
\code{\link{set_thumbnail}()},
\code{\link{set_vanity_url}()},
\code{\link{swap_vanity_url}()},
\code{\link{verify_content_name}()}
}
\concept{content functions}
