% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy.R
\name{dashboard_url}
\alias{dashboard_url}
\title{Build a Dashboard URL from a Content Item}
\usage{
dashboard_url(content, pane = "")
}
\arguments{
\item{content}{\link{Content} A Content object}

\item{pane}{character The pane in the dashboard to link to}
}
\value{
character The dashboard URL for the content provided
}
\description{
Returns the URL for the content dashboard (opened to the selected pane).
}
\seealso{
Other content functions: 
\code{\link{acl_add_user}()},
\code{\link{content_item}()},
\code{\link{content_title}()},
\code{\link{dashboard_url_chr}()},
\code{\link{get_acl_user}()},
\code{\link{get_environment}()},
\code{\link{get_image}()},
\code{\link{get_jobs}()},
\code{\link{get_vanity_url}()},
\code{\link{set_image_path}()},
\code{\link{set_run_as}()},
\code{\link{set_vanity_url}()},
\code{\link{swap_vanity_url}()}
}
\concept{content functions}
