% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy.R
\name{deploy}
\alias{deploy}
\title{Deploy a bundle}
\usage{
deploy(connect, bundle, name = random_name(), title = name, guid = NULL, ...)
}
\arguments{
\item{connect}{A Connect object}

\item{bundle}{A Bundle object}

\item{name}{The unique name for the content on the server}

\item{title}{optional The title to be used for the content on the server}

\item{guid}{optional The GUID if the content already exists on the server}

\item{...}{Additional arguments passed along to the content creation}
}
\value{
Task A task object
}
\description{
Deploy a bundle
}
\seealso{
Other deployment functions: 
\code{\link{bundle_dir}()},
\code{\link{bundle_path}()},
\code{\link{bundle_static}()},
\code{\link{download_bundle}()},
\code{\link{poll_task}()}
}
\concept{deployment functions}
