% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.R
\name{deploy_repo}
\alias{deploy_repo}
\title{Deploy a Git Repository}
\usage{
deploy_repo(
  client,
  repository,
  branch,
  subdirectory,
  name = create_random_name(),
  title = name,
  ...
)
}
\arguments{
\item{client}{A Connect R6 object}

\item{repository}{The git repository to deploy}

\item{branch}{The git branch to deploy}

\item{subdirectory}{The subdirectory to deploy (must contain a \code{manifest.json})}

\item{name}{The "name" / unique identifier for the content. Defaults to a random character string}

\item{title}{The "title" of the content}

\item{...}{Additional options for defining / specifying content attributes}
}
\value{
A ContentTask object, for use with \code{poll_task()} (if you want to follow the logs)
}
\description{
\lifecycle{experimental} Deploy a git repository directly to RStudio Connect,
using RStudio Connect's "pull-based" "git-polling" method of deployment.
}
\seealso{
connectapi::poll_task, connectapi::repo_check_branches, connectapi::repo_check_manifest_dirs

Other content functions: 
\code{\link{acl_add_user}()},
\code{\link{content_item}()},
\code{\link{content_title}()},
\code{\link{create_random_name}()},
\code{\link{dashboard_url_chr}()},
\code{\link{dashboard_url}()},
\code{\link{delete_vanity_url}()},
\code{\link{get_acl_user}()},
\code{\link{get_bundles}()},
\code{\link{get_environment}()},
\code{\link{get_image}()},
\code{\link{get_jobs}()},
\code{\link{get_vanity_url}()},
\code{\link{git}},
\code{\link{set_image_path}()},
\code{\link{set_run_as}()},
\code{\link{set_vanity_url}()},
\code{\link{swap_vanity_url}()},
\code{\link{verify_content_name}()}
}
\concept{content functions}
