% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy.R
\name{download_bundle}
\alias{download_bundle}
\title{Download a Bundle from Deployed Connect Content}
\usage{
download_bundle(
  content,
  filename = fs::file_temp(pattern = "bundle", ext = ".tar.gz")
)
}
\arguments{
\item{content}{A Content object}

\item{filename}{The output bundle path}
}
\value{
Bundle A bundle object
}
\description{
Downloads a Content item's active bundle
}
\seealso{
Other deployment functions: 
\code{\link{bundle_dir}()},
\code{\link{bundle_path}()},
\code{\link{bundle_static}()},
\code{\link{deploy}()},
\code{\link{poll_task}()}
}
\concept{deployment functions}
