% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content.R
\name{get_environment}
\alias{get_environment}
\alias{set_environment_new}
\alias{set_environment_remove}
\title{Manage Environment Variables}
\usage{
get_environment(content)

set_environment_new(env, ...)

set_environment_remove(env, ...)
}
\arguments{
\item{content}{An R6 Content object as returned by \code{content_item()}}

\item{env}{An R6 Environment object as returned by \code{get_environment()}}

\item{...}{name = value pairs of environment variable names and values}
}
\description{
\lifecycle{experimental} Manage Environment Variables for a piece of content.
}
\details{
\code{get_environment()} returns an Environment object for use with "setter" methods

\code{set_environment_new()} sets new environment values (either creating new
values or updating existing)

\code{set_environment_remove()} removes existing environment variables
}
\seealso{
Other content functions: 
\code{\link{acl_add_user}()},
\code{\link{content_item}()},
\code{\link{content_title}()},
\code{\link{create_random_name}()},
\code{\link{dashboard_url_chr}()},
\code{\link{dashboard_url}()},
\code{\link{get_acl_user}()},
\code{\link{get_bundles}()},
\code{\link{get_image}()},
\code{\link{get_jobs}()},
\code{\link{get_vanity_url}()},
\code{\link{set_image_path}()},
\code{\link{set_run_as}()},
\code{\link{set_vanity_url}()},
\code{\link{swap_vanity_url}()},
\code{\link{verify_content_name}()}
}
\concept{content functions}
