% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups.R
\name{get_group_members}
\alias{get_group_members}
\title{Get users within a specific group}
\usage{
get_group_members(src, guid)
}
\arguments{
\item{src}{A Connect client object}

\item{guid}{A group GUID identifier}
}
\value{
A tibble with the following columns:
\itemize{
\item \code{email}: The user's email
\item \code{username}: The user's username
\item \code{first_name}: The user's first name
\item \code{last_name}: The user's last name
\item \code{user_role}: The user's role. It may have a value of administrator,
publisher or viewer.
\item \code{created_time}: The timestamp (in RFC3339 format) when the user
was created in the Posit Connect server
\item \code{updated_time}: The timestamp (in RFC3339 format) when the user
was last updated in the Posit Connect server
\item \code{active_time}: The timestamp (in RFC3339 format) when the user
was last active on the Posit Connect server
\item \code{confirmed}: When false, the created user must confirm their
account through an email. This feature is unique to password
authentication.
\item \code{locked}: Whether or not the user is locked
\item \code{guid}: The user's GUID, or unique identifier, in UUID RFC4122 format
}
}
\description{
Get users within a specific group
}
\details{
Please see https://docs.posit.co/connect/api/#get-/v1/groups/-group_guid-/members
for more information.
}
\examples{
\dontrun{
library(connectapi)
client <- connect()

# get the first 20 groups
groups <- get_groups(client)

group_guid <- groups$guid[1]

get_group_members(client, guid = group_guid)
}

}
\seealso{
Other groups functions: 
\code{\link{get_group_content}()},
\code{\link{get_groups}()}
}
\concept{groups functions}
