% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy.R
\name{get_image}
\alias{get_image}
\title{Get the Image}
\usage{
get_image(content, path)
}
\arguments{
\item{content}{A content object}

\item{path}{The path to the image on disk}
}
\description{
Get the Image
}
\seealso{
Other content: \code{\link{content_item}},
  \code{\link{get_vanity_url}},
  \code{\link{set_image_path}},
  \code{\link{set_vanity_url}},
  \code{\link{swap_vanity_url}}
}
\concept{content}
