% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy.R
\name{get_image}
\alias{get_image}
\alias{delete_image}
\alias{has_image}
\title{Get the Image}
\usage{
get_image(content, path = NULL)

delete_image(content, path = NULL)

has_image(content)
}
\arguments{
\item{content}{A content object}

\item{path}{optional. The path to the image on disk}
}
\description{
\lifecycle{experimental}
\code{get_image} saves the content image to the given path (default: temp file).
\code{delete_image} removes the image (optionally saving to the given path)
\code{has_image} returns whether the content has an image
}
\seealso{
Other deployment functions: 
\code{\link{bundle_dir}()},
\code{\link{bundle_path}()},
\code{\link{bundle_static}()},
\code{\link{deploy}()},
\code{\link{download_bundle}()},
\code{\link{poll_task}()}

Other content: 
\code{\link{swap_vanity_url}()}

Other content: 
\code{\link{swap_vanity_url}()}
}
\concept{content}
\concept{deployment functions}
