% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy.R
\name{get_image}
\alias{get_image}
\alias{delete_image}
\alias{has_image}
\title{Get the Image}
\usage{
get_image(content, path = NULL)

delete_image(content, path = NULL)

has_image(content)
}
\arguments{
\item{content}{A content object}

\item{path}{optional. The path to the image on disk}
}
\description{
\lifecycle{experimental}
\code{get_image} saves the content image to the given path (default: temp file).
\code{delete_image} removes the image (optionally saving to the given path)
\code{has_image} returns whether the content has an image
}
\seealso{
Other content functions: 
\code{\link{acl_add_user}()},
\code{\link{content_item}()},
\code{\link{content_title}()},
\code{\link{dashboard_url_chr}()},
\code{\link{dashboard_url}()},
\code{\link{get_acl_user}()},
\code{\link{get_environment}()},
\code{\link{get_jobs}()},
\code{\link{get_vanity_url}()},
\code{\link{set_image_path}()},
\code{\link{set_run_as}()},
\code{\link{set_vanity_url}()},
\code{\link{swap_vanity_url}()}
}
\concept{content functions}
