% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_procs}
\alias{get_procs}
\title{Get Real-Time Process Data}
\usage{
get_procs(src)
}
\arguments{
\item{src}{The source object}
}
\value{
A tibble with the following columns:
\itemize{
\item{\strong{pid}}{The PID of the current process}
\item{\strong{appId}}{The application ID}
\item{\strong{appGuid}}{The application GUID}
\item{\strong{appName}}{The application name}
\item{\strong{appUrl}}{The application URL}
\item{\strong{appRunAs}}{The application RunAs user}
\item{\strong{type}}{The type of process}
\item{\strong{cpuCurrent}}{The current CPU usage}
\item{\strong{cpuTotal}}{The total CPU usage}
\item{\strong{ram}}{The current RAM usage}
}
}
\description{
\lifecycle{experimental}
This returns real-time process data from the RStudio Connect API. It requires
administrator privileges to use. NOTE that this only returns data for the
server that responds to the request (i.e. in a Highly Available cluster)
}
