% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runtime-caches.R
\name{get_runtime_caches}
\alias{get_runtime_caches}
\title{Get runtime caches}
\usage{
get_runtime_caches(client)
}
\arguments{
\item{client}{A \code{Connect} object.}
}
\value{
A tibble of runtime caches on the server, showing \code{language},
\code{version} and \code{image_name}. For Connect servers not using off-host execution,
\code{image_name} is \code{"Local"}.
}
\description{
View the runtime caches on a Connect server.
Requires administrator privileges.
}
\examples{
\dontrun{
client <- connect()
get_runtime_caches(client)
}

}
\seealso{
Other server management functions: 
\code{\link{delete_runtime_cache}()}
}
\concept{server management functions}
