% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy.R
\name{get_thumbnail}
\alias{get_thumbnail}
\alias{get_image}
\alias{delete_thumbnail}
\alias{delete_image}
\alias{has_thumbnail}
\alias{has_image}
\title{Get the Content Thumbnail}
\usage{
get_thumbnail(content, path = NULL)

get_image(content, path = NULL)

delete_thumbnail(content, path = NULL)

delete_image(content, path = NULL)

has_thumbnail(content)

has_image(content)
}
\arguments{
\item{content}{A content object}

\item{path}{optional. The path to the image on disk}
}
\description{
\lifecycle{experimental}
\code{get_thumbnail} saves the content thumbnail to the given path (default: temp file).
\code{delete_thumbnail} removes the thumbnail (optionally saving to the given path)
\code{has_thumbnail} returns whether the content has an thumbnail
}
\seealso{
Other content functions: 
\code{\link{content_delete}()},
\code{\link{content_item}()},
\code{\link{content_title}()},
\code{\link{content_update}()},
\code{\link{create_random_name}()},
\code{\link{dashboard_url}()},
\code{\link{dashboard_url_chr}()},
\code{\link{delete_vanity_url}()},
\code{\link{deploy_repo}()},
\code{\link{get_bundles}()},
\code{\link{get_environment}()},
\code{\link{get_jobs}()},
\code{\link{get_vanity_url}()},
\code{\link{git}},
\code{\link{permissions}},
\code{\link{set_run_as}()},
\code{\link{set_thumbnail}()},
\code{\link{set_vanity_url}()},
\code{\link{swap_vanity_url}()},
\code{\link{verify_content_name}()}
}
\concept{content functions}
