% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy.R
\name{get_vanity_url}
\alias{get_vanity_url}
\title{Get the Vanity URL}
\usage{
get_vanity_url(content)
}
\arguments{
\item{content}{A Content object}
}
\description{
\lifecycle{experimental} Gets the Vanity URL for a piece of content.
}
\seealso{
Other content functions: 
\code{\link{content_item}()},
\code{\link{content_title}()},
\code{\link{dashboard_url_chr}()},
\code{\link{dashboard_url}()},
\code{\link{get_acl_user}()},
\code{\link{get_environment}()},
\code{\link{set_image_path}()},
\code{\link{set_run_as}()},
\code{\link{set_vanity_url}()}
}
\concept{content functions}
