% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy.R
\name{get_vanity_url}
\alias{get_vanity_url}
\title{Get the Vanity URL}
\usage{
get_vanity_url(content)
}
\arguments{
\item{content}{A Content object}
}
\description{
Gets the Vanity URL for a piece of content.
}
\seealso{
Other content: \code{\link{content_item}},
  \code{\link{get_image}}, \code{\link{set_image_path}},
  \code{\link{set_vanity_url}}
}
\concept{content}
