% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schedule.R
\name{get_variant_schedule}
\alias{get_variant_schedule}
\title{Get a Variant Schedule}
\usage{
get_variant_schedule(variant)
}
\arguments{
\item{variant}{A Variant object, as returned by \code{get_variant()} or \code{get_variant_default()}}
}
\value{
A VariantSchedule object
}
\description{
\lifecycle{experimental} Gets the schedule associated with a Variant.
}
\seealso{
Other schedule functions: 
\code{\link{set_schedule}()}
}
\concept{schedule functions}
