% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remote.R
\name{groups_create_remote}
\alias{groups_create_remote}
\title{Create a Remote Group}
\usage{
groups_create_remote(connect, prefix, expect = 1, check = TRUE)
}
\arguments{
\item{connect}{A R6 Connect object}

\item{prefix}{character. The prefix of the group name to search for}

\item{expect}{number. The number of responses to expect for this search}

\item{check}{boolean. Whether to check for local existence first}
}
\value{
The results of creating the groups
}
\description{
Create a Remote Group
}
