% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content.R
\name{get_jobs}
\alias{get_jobs}
\alias{get_job}
\title{Get Jobs}
\usage{
get_jobs(content)

get_job(content, key)
}
\description{
\lifecycle{experimental} Retrieve details about jobs associated with a \code{content_item}.
"Jobs" in RStudio Connect are content executions
}
\seealso{
Other content functions: 
\code{\link{content_item}()},
\code{\link{content_title}()},
\code{\link{dashboard_url_chr}()},
\code{\link{dashboard_url}()},
\code{\link{get_acl_user}()},
\code{\link{get_environment}()},
\code{\link{get_vanity_url}()},
\code{\link{set_image_path}()},
\code{\link{set_run_as}()},
\code{\link{set_vanity_url}()}
}
\concept{content functions}
