% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy.R
\name{poll_task}
\alias{poll_task}
\title{Poll Task}
\usage{
poll_task(task, wait = 1)
}
\description{
Poll Task
}
\seealso{
Other deploy: \code{\link{bundle_dir}},
  \code{\link{bundle_path}}, \code{\link{deploy}}
}
\concept{deploy}
