% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy.R
\name{poll_task}
\alias{poll_task}
\title{Poll Task}
\usage{
poll_task(task, wait = 1)
}
\arguments{
\item{task}{A Task object}

\item{wait}{The interval to wait between polling}
}
\value{
Task The Task object that was input
}
\description{
Polls a task, waiting for information about a deployment
}
\seealso{
Other deployment functions: 
\code{\link{bundle_dir}()},
\code{\link{bundle_path}()},
\code{\link{bundle_static}()},
\code{\link{deploy}()},
\code{\link{download_bundle}()},
\code{\link{get_image}()}
}
\concept{deployment functions}
