% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy.R
\name{set_image_path}
\alias{set_image_path}
\alias{set_image_url}
\alias{set_image_webshot}
\title{Set the Image from a Path}
\usage{
set_image_path(content, path)

set_image_url(content, url)

set_image_webshot(content, ...)
}
\arguments{
\item{content}{A content object}

\item{path}{The path to an image on disk}

\item{url}{The url for an image}

\item{...}{Additional arguments passed on to methods}
}
\description{
Set the Image from a Path
}
\seealso{
Other content: \code{\link{set_vanity_url}}
}
\concept{content}
