% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy.R
\name{set_image_path}
\alias{set_image_path}
\alias{set_image_url}
\title{Set the Image from a Path}
\usage{
set_image_path(content, path)

set_image_url(content, url)
}
\arguments{
\item{content}{A content object}

\item{path}{The path to an image on disk}

\item{url}{The url for an image}
}
\description{
\lifecycle{experimental}
}
\seealso{
Other content functions: 
\code{\link{acl_add_user}()},
\code{\link{content_item}()},
\code{\link{content_title}()},
\code{\link{create_random_name}()},
\code{\link{dashboard_url_chr}()},
\code{\link{dashboard_url}()},
\code{\link{get_acl_user}()},
\code{\link{get_environment}()},
\code{\link{get_image}()},
\code{\link{get_jobs}()},
\code{\link{get_vanity_url}()},
\code{\link{set_run_as}()},
\code{\link{set_vanity_url}()},
\code{\link{swap_vanity_url}()},
\code{\link{verify_content_name}()}
}
\concept{content functions}
