% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schedule.R
\name{set_schedule}
\alias{set_schedule}
\alias{set_schedule_remove}
\title{Set a Schedule}
\usage{
set_schedule(.schedule, ...)

set_schedule_remove(.schedule)
}
\arguments{
\item{.schedule}{A schedule object. As returned by \code{get_variant_schedule()}}

\item{type}{The type of schedule. One of "hour", "minute", "day", "weekday",
"week", "semimonth", "dayofmonth", "dayweekofmonth", "year"}

\item{email}{Whether to send emails on this schedule}

\item{activate}{Whether to publish the output of this schedule}

\item{schedule}{A JSON blob (as a string) describing the schedule. See "More Details"}

\item{start_time}{The start time of the schedule}

\item{next_run}{The next run of the schedule (set to \code{start_time} in most cases)}

\item{n}{The "number of" iterations}

\item{day}{The day of the week (0-6) or day of the month (0-31)}

\item{days}{The days of the week (0-6)}

\item{week}{The week of the month (0-5)}

\item{first}{\link{logical} Whether to execute on the 1st and 15th (TRUE) or 14th and last (FALSE)}
}
\value{
An updated Schedule object
}
\description{
\lifecycle{experimental} Sets the schedule for a given Variant. Requires a
\code{Schedule} object (as returned by \code{get_variant_schedule()})
}
\details{
\itemize{
\item \code{set_schedule()} is a somewhat raw interface to RStudio Connect's \code{schedule} API
\item \verb{set_schedule_*()} functions provide handy wrappers around \code{set_schedule()}
\item \code{set_schedule_remove()} removes a schedule / un-schedules a variant
}
}
\seealso{
Other schedule functions: 
\code{\link{get_variant_schedule}()}
}
\concept{schedule functions}
