% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy.R
\name{set_vanity_url}
\alias{set_vanity_url}
\title{Set the Vanity URL}
\usage{
set_vanity_url(content, url)
}
\arguments{
\item{content}{A Content object}

\item{url}{The path component of the URL}
}
\value{
An updated Content object
}
\description{
Sets the Vanity URL for a piece of content.
}
\examples{
\dontrun{
bnd <- bundle_dir("~/my/directory")
connect() \%>\% 
  deploy(bnd) \%>\% 
  set_vanity_url("a/vanity/url")
}

}
\seealso{
Other content: \code{\link{content_item}},
  \code{\link{get_vanity_url}},
  \code{\link{set_image_path}}
}
\concept{content}
