% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy.R
\name{swap_vanity_url}
\alias{swap_vanity_url}
\title{Swap the Vanity URL}
\usage{
swap_vanity_url(from_content, to_content)
}
\arguments{
\item{from_content}{A Content object}

\item{to_content}{A Content object}
}
\description{
\lifecycle{experimental}
Swaps the Vanity URLs between two pieces of content
}
\seealso{
Other content: \code{\link{content_item}},
  \code{\link{get_image}}, \code{\link{get_vanity_url}},
  \code{\link{set_image_path}},
  \code{\link{set_vanity_url}}
}
\concept{content}
