% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy.R
\name{swap_vanity_url}
\alias{swap_vanity_url}
\title{Swap the Vanity URL}
\usage{
swap_vanity_url(from_content, to_content)
}
\arguments{
\item{from_content}{A Content object}

\item{to_content}{A Content object}
}
\description{
Swaps the Vanity URLs between two pieces of content
}
\seealso{
Other content functions: 
\code{\link{content_delete}()},
\code{\link{content_item}()},
\code{\link{content_title}()},
\code{\link{content_update}()},
\code{\link{create_random_name}()},
\code{\link{dashboard_url}()},
\code{\link{dashboard_url_chr}()},
\code{\link{delete_thumbnail}()},
\code{\link{delete_vanity_url}()},
\code{\link{deploy_repo}()},
\code{\link{get_bundles}()},
\code{\link{get_environment}()},
\code{\link{get_jobs}()},
\code{\link{get_thumbnail}()},
\code{\link{get_vanity_url}()},
\code{\link{git}},
\code{\link{has_thumbnail}()},
\code{\link{permissions}},
\code{\link{set_run_as}()},
\code{\link{set_thumbnail}()},
\code{\link{set_vanity_url}()},
\code{\link{verify_content_name}()}
}
\concept{content functions}
