% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag_page.R
\name{tag_page}
\alias{tag_page}
\title{Generate a landing page for a specific RStudio Connect tag}
\usage{
tag_page(connect, tag, description = NULL)
}
\arguments{
\item{connect}{A Connect object}

\item{tag}{The name of the targeted tag. NOTE: tag names are not necessarily unique}

\item{description}{A description of the tag, placed at the top of the landing
page}
}
\value{
A list with LANDING_PAGE which is the path to the html file and APPS
  which is a list containing information on the apps and screenshots, useful
  if you'd rather use your own template
}
\description{
\lifecycle{experimental}
Note that in order to show a content item in a page, you must have "Viewer"
priveleges to that content.
}
