% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags.R
\name{get_tags}
\alias{get_tags}
\alias{get_tag_data}
\alias{create_tag}
\alias{create_tag_tree}
\alias{set_content_tags}
\alias{filter_tag_tree}
\alias{get_content_tags}
\alias{delete_tag}
\title{Get all Tags on the server}
\usage{
get_tags(src)

get_tag_data(src)

create_tag(src, name, parent = NULL)

create_tag_tree(src, ...)

set_content_tags(content, ...)

filter_tag_tree(tags, ids)

get_content_tags(content)

delete_tag(src, tag)
}
\arguments{
\item{src}{The source object}

\item{name}{The name of the tag to create}

\item{parent}{optional. A \code{connect_tag_tree} object (as returned by \code{get_tags()}) pointed at the parent tag}

\item{...}{Additional arguments}

\item{content}{An R6 Content object, as returned by \code{content_item()}}

\item{tags}{A \code{connect_tag_tree} object (as returned by \code{get_tags()})}

\item{ids}{A list of \code{id}s to filter the tag tree by}

\item{tag}{A \code{connect_tag_tree} object (as returned by \code{get_tags()})}
}
\description{
\lifecycle{experimental} Get a tibble of all tags
}
