% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy.R
\docType{data}
\name{Task}
\alias{Task}
\title{Task}
\format{An object of class \code{R6ClassGenerator} of length 25.}
\usage{
Task
}
\description{
An R6 class that represents a Task
}
\seealso{
Other R6 classes: \code{\link{Bundle}},
  \code{\link{Content}}, \code{\link{RStudioConnect}},
  \code{\link{Vanity}}
}
\concept{R6 classes}
\keyword{datasets}
