% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy.R
\docType{data}
\name{Task}
\alias{Task}
\title{Task}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Task
}
\description{
An S6 class that represents a Task
}
\seealso{
Other deploy: \code{\link{Bundle}},
  \code{\link{bundle_dir}}, \code{\link{bundle_path}},
  \code{\link{deploy}}
}
\concept{deploy}
\keyword{datasets}
