% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazy.R
\name{tbl_connect}
\alias{tbl_connect}
\title{Connect Tibble}
\usage{
tbl_connect(
  src,
  from = c("users", "groups", "content", "usage_shiny", "usage_static", "audit_logs"),
  ...
)
}
\arguments{
\item{src}{The source object}

\item{from}{The type of tibble}

\item{...}{Additional arguments that are not yet implemented}
}
\value{
A \code{tbl_connect} object
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
A lazy tibble that automatically pages through API requests when \code{collect}ed.
}
