% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazy.R
\name{tbl_connect}
\alias{tbl_connect}
\title{Connect Tibble}
\usage{
tbl_connect(src, from = c("users", "content", "shiny_usage",
  "content_visits"), ...)
}
\arguments{
\item{src}{The source object}

\item{from}{The type of tibble}

\item{...}{Additional arguments that are not yet implemented}
}
\value{
A `tbl_connect` object
}
\description{
\lifecycle{experimental}
A lazy tibble that automatically pages through API requests when `collect`ed.
}
