% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remote.R
\name{users_create_remote}
\alias{users_create_remote}
\title{Create a Remote User}
\usage{
users_create_remote(src, prefix, expect = 1, check = TRUE)
}
\arguments{
\item{src}{A R6 Connect object}

\item{prefix}{character. The prefix of the user name to search for}

\item{expect}{number. The number of responses to expect for this search}

\item{check}{boolean. Whether to check for local existence first}
}
\value{
The results of creating the users
}
\description{
Create a Remote User
}
