% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy.R
\docType{data}
\name{Vanity}
\alias{Vanity}
\title{Vanity}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Vanity
}
\description{
An R6 class that represents a Vanity URL
}
\seealso{
Other deployR6: \code{\link{Bundle}},
  \code{\link{Content}}, \code{\link{Task}}
}
\concept{deployR6}
\keyword{datasets}
