% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audits.R
\name{vanity_is_available}
\alias{vanity_is_available}
\title{Check to see if a vanity URL is currently in use}
\usage{
vanity_is_available(connect, vanity)
}
\arguments{
\item{connect}{A Connect R6 object}

\item{vanity}{string of the vanity URL to check}
}
\value{
logical indicating if the vanity URL is available.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\seealso{
Other audit functions: 
\code{\link{audit_access_open}()},
\code{\link{audit_r_versions}()},
\code{\link{audit_runas}()}
}
\concept{audit functions}
