% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant.R
\name{Variant}
\alias{Variant}
\title{Variant}
\description{
An R6 class that represents a Variant
}
\seealso{
Other R6 classes: 
\code{\link{Bundle}},
\code{\link{ContentTask}},
\code{\link{Content}},
\code{\link{Environment}},
\code{\link{RStudioConnect}},
\code{\link{Task}},
\code{\link{Vanity}},
\code{\link{VariantSchedule}},
\code{\link{VariantTask}}
}
\concept{R6 classes}
\section{Super class}{
\code{\link[connectapi:Content]{connectapi::Content}} -> \code{Variant}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-get_variant}{\code{Variant$get_variant()}}
\item \href{#method-get_variant_remote}{\code{Variant$get_variant_remote()}}
\item \href{#method-new}{\code{Variant$new()}}
\item \href{#method-send_mail}{\code{Variant$send_mail()}}
\item \href{#method-get_schedule}{\code{Variant$get_schedule()}}
\item \href{#method-get_schedule_remote}{\code{Variant$get_schedule_remote()}}
\item \href{#method-get_subscribers}{\code{Variant$get_subscribers()}}
\item \href{#method-remove_subscriber}{\code{Variant$remove_subscriber()}}
\item \href{#method-add_subscribers}{\code{Variant$add_subscribers()}}
\item \href{#method-render}{\code{Variant$render()}}
\item \href{#method-renderings}{\code{Variant$renderings()}}
\item \href{#method-update_variant}{\code{Variant$update_variant()}}
\item \href{#method-jobs}{\code{Variant$jobs()}}
\item \href{#method-job}{\code{Variant$job()}}
\item \href{#method-get_url}{\code{Variant$get_url()}}
\item \href{#method-get_url_rev}{\code{Variant$get_url_rev()}}
\item \href{#method-get_dashboard_url}{\code{Variant$get_dashboard_url()}}
\item \href{#method-print}{\code{Variant$print()}}
\item \href{#method-clone}{\code{Variant$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="connectapi" data-topic="Content" data-id="deploy">}\href{../../connectapi/html/Content.html#method-deploy}{\code{connectapi::Content$deploy()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="connectapi" data-topic="Content" data-id="environment">}\href{../../connectapi/html/Content.html#method-environment}{\code{connectapi::Content$environment()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="connectapi" data-topic="Content" data-id="environment_all">}\href{../../connectapi/html/Content.html#method-environment_all}{\code{connectapi::Content$environment_all()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="connectapi" data-topic="Content" data-id="environment_set">}\href{../../connectapi/html/Content.html#method-environment_set}{\code{connectapi::Content$environment_set()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="connectapi" data-topic="Content" data-id="get_bundles">}\href{../../connectapi/html/Content.html#method-get_bundles}{\code{connectapi::Content$get_bundles()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="connectapi" data-topic="Content" data-id="get_connect">}\href{../../connectapi/html/Content.html#method-get_connect}{\code{connectapi::Content$get_connect()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="connectapi" data-topic="Content" data-id="get_content">}\href{../../connectapi/html/Content.html#method-get_content}{\code{connectapi::Content$get_content()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="connectapi" data-topic="Content" data-id="get_content_remote">}\href{../../connectapi/html/Content.html#method-get_content_remote}{\code{connectapi::Content$get_content_remote()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="connectapi" data-topic="Content" data-id="get_job">}\href{../../connectapi/html/Content.html#method-get_job}{\code{connectapi::Content$get_job()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="connectapi" data-topic="Content" data-id="get_jobs">}\href{../../connectapi/html/Content.html#method-get_jobs}{\code{connectapi::Content$get_jobs()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="connectapi" data-topic="Content" data-id="internal_content">}\href{../../connectapi/html/Content.html#method-internal_content}{\code{connectapi::Content$internal_content()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="connectapi" data-topic="Content" data-id="permissions">}\href{../../connectapi/html/Content.html#method-permissions}{\code{connectapi::Content$permissions()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="connectapi" data-topic="Content" data-id="permissions_add">}\href{../../connectapi/html/Content.html#method-permissions_add}{\code{connectapi::Content$permissions_add()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="connectapi" data-topic="Content" data-id="permissions_delete">}\href{../../connectapi/html/Content.html#method-permissions_delete}{\code{connectapi::Content$permissions_delete()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="connectapi" data-topic="Content" data-id="permissions_update">}\href{../../connectapi/html/Content.html#method-permissions_update}{\code{connectapi::Content$permissions_update()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="connectapi" data-topic="Content" data-id="repo_enable">}\href{../../connectapi/html/Content.html#method-repo_enable}{\code{connectapi::Content$repo_enable()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="connectapi" data-topic="Content" data-id="repo_set">}\href{../../connectapi/html/Content.html#method-repo_set}{\code{connectapi::Content$repo_set()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="connectapi" data-topic="Content" data-id="runas">}\href{../../connectapi/html/Content.html#method-runas}{\code{connectapi::Content$runas()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="connectapi" data-topic="Content" data-id="tag_delete">}\href{../../connectapi/html/Content.html#method-tag_delete}{\code{connectapi::Content$tag_delete()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="connectapi" data-topic="Content" data-id="tag_set">}\href{../../connectapi/html/Content.html#method-tag_set}{\code{connectapi::Content$tag_set()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="connectapi" data-topic="Content" data-id="tags">}\href{../../connectapi/html/Content.html#method-tags}{\code{connectapi::Content$tags()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="connectapi" data-topic="Content" data-id="update">}\href{../../connectapi/html/Content.html#method-update}{\code{connectapi::Content$update()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="connectapi" data-topic="Content" data-id="variants">}\href{../../connectapi/html/Content.html#method-variants}{\code{connectapi::Content$variants()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_variant"></a>}}
\if{latex}{\out{\hypertarget{method-get_variant}{}}}
\subsection{Method \code{get_variant()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Variant$get_variant()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_variant_remote"></a>}}
\if{latex}{\out{\hypertarget{method-get_variant_remote}{}}}
\subsection{Method \code{get_variant_remote()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Variant$get_variant_remote()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Variant$new(connect, content, key)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-send_mail"></a>}}
\if{latex}{\out{\hypertarget{method-send_mail}{}}}
\subsection{Method \code{send_mail()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Variant$send_mail(to = c("me", "collaborators", "collaborators_viewers"))}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_schedule"></a>}}
\if{latex}{\out{\hypertarget{method-get_schedule}{}}}
\subsection{Method \code{get_schedule()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Variant$get_schedule()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_schedule_remote"></a>}}
\if{latex}{\out{\hypertarget{method-get_schedule_remote}{}}}
\subsection{Method \code{get_schedule_remote()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Variant$get_schedule_remote()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_subscribers"></a>}}
\if{latex}{\out{\hypertarget{method-get_subscribers}{}}}
\subsection{Method \code{get_subscribers()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Variant$get_subscribers()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-remove_subscriber"></a>}}
\if{latex}{\out{\hypertarget{method-remove_subscriber}{}}}
\subsection{Method \code{remove_subscriber()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Variant$remove_subscriber(guid)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_subscribers"></a>}}
\if{latex}{\out{\hypertarget{method-add_subscribers}{}}}
\subsection{Method \code{add_subscribers()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Variant$add_subscribers(guids)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-render"></a>}}
\if{latex}{\out{\hypertarget{method-render}{}}}
\subsection{Method \code{render()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Variant$render()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-renderings"></a>}}
\if{latex}{\out{\hypertarget{method-renderings}{}}}
\subsection{Method \code{renderings()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Variant$renderings()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_variant"></a>}}
\if{latex}{\out{\hypertarget{method-update_variant}{}}}
\subsection{Method \code{update_variant()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Variant$update_variant(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-jobs"></a>}}
\if{latex}{\out{\hypertarget{method-jobs}{}}}
\subsection{Method \code{jobs()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Variant$jobs()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-job"></a>}}
\if{latex}{\out{\hypertarget{method-job}{}}}
\subsection{Method \code{job()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Variant$job(key)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_url"></a>}}
\if{latex}{\out{\hypertarget{method-get_url}{}}}
\subsection{Method \code{get_url()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Variant$get_url()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_url_rev"></a>}}
\if{latex}{\out{\hypertarget{method-get_url_rev}{}}}
\subsection{Method \code{get_url_rev()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Variant$get_url_rev(rev)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_dashboard_url"></a>}}
\if{latex}{\out{\hypertarget{method-get_dashboard_url}{}}}
\subsection{Method \code{get_dashboard_url()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Variant$get_dashboard_url(pane = "access")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Variant$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Variant$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
