% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schedule.R
\name{VariantSchedule}
\alias{VariantSchedule}
\title{VariantSchedule}
\description{
An R6 class that represents a Schedule
}
\section{Super classes}{
\code{\link[connectapi:Content]{connectapi::Content}} -> \code{\link[connectapi:Variant]{connectapi::Variant}} -> \code{VariantSchedule}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{VariantSchedule$new()}}
\item \href{#method-GET}{\code{VariantSchedule$GET()}}
\item \href{#method-POST}{\code{VariantSchedule$POST()}}
\item \href{#method-DELETE}{\code{VariantSchedule$DELETE()}}
\item \href{#method-set_schedule}{\code{VariantSchedule$set_schedule()}}
\item \href{#method-is_empty}{\code{VariantSchedule$is_empty()}}
\item \href{#method-print}{\code{VariantSchedule$print()}}
\item \href{#method-get_schedule}{\code{VariantSchedule$get_schedule()}}
\item \href{#method-get_schedule_remote}{\code{VariantSchedule$get_schedule_remote()}}
\item \href{#method-describe_schedule}{\code{VariantSchedule$describe_schedule()}}
\item \href{#method-clone}{\code{VariantSchedule$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="connectapi" data-topic="Content" data-id="environment">}\href{../../connectapi/html/Content.html#method-environment}{\code{connectapi::Content$environment()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="connectapi" data-topic="Content" data-id="environment_set">}\href{../../connectapi/html/Content.html#method-environment_set}{\code{connectapi::Content$environment_set()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="connectapi" data-topic="Content" data-id="get_bundles">}\href{../../connectapi/html/Content.html#method-get_bundles}{\code{connectapi::Content$get_bundles()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="connectapi" data-topic="Content" data-id="get_connect">}\href{../../connectapi/html/Content.html#method-get_connect}{\code{connectapi::Content$get_connect()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="connectapi" data-topic="Content" data-id="get_content">}\href{../../connectapi/html/Content.html#method-get_content}{\code{connectapi::Content$get_content()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="connectapi" data-topic="Content" data-id="get_content_remote">}\href{../../connectapi/html/Content.html#method-get_content_remote}{\code{connectapi::Content$get_content_remote()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="connectapi" data-topic="Content" data-id="get_job">}\href{../../connectapi/html/Content.html#method-get_job}{\code{connectapi::Content$get_job()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="connectapi" data-topic="Content" data-id="get_jobs">}\href{../../connectapi/html/Content.html#method-get_jobs}{\code{connectapi::Content$get_jobs()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="connectapi" data-topic="Content" data-id="runas">}\href{../../connectapi/html/Content.html#method-runas}{\code{connectapi::Content$runas()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="connectapi" data-topic="Content" data-id="tag_delete">}\href{../../connectapi/html/Content.html#method-tag_delete}{\code{connectapi::Content$tag_delete()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="connectapi" data-topic="Content" data-id="tag_set">}\href{../../connectapi/html/Content.html#method-tag_set}{\code{connectapi::Content$tag_set()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="connectapi" data-topic="Content" data-id="tags">}\href{../../connectapi/html/Content.html#method-tags}{\code{connectapi::Content$tags()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="connectapi" data-topic="Content" data-id="update">}\href{../../connectapi/html/Content.html#method-update}{\code{connectapi::Content$update()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="connectapi" data-topic="Content" data-id="variants">}\href{../../connectapi/html/Content.html#method-variants}{\code{connectapi::Content$variants()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="connectapi" data-topic="Variant" data-id="add_subscribers">}\href{../../connectapi/html/Variant.html#method-add_subscribers}{\code{connectapi::Variant$add_subscribers()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="connectapi" data-topic="Variant" data-id="get_dashboard_url">}\href{../../connectapi/html/Variant.html#method-get_dashboard_url}{\code{connectapi::Variant$get_dashboard_url()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="connectapi" data-topic="Variant" data-id="get_subscribers">}\href{../../connectapi/html/Variant.html#method-get_subscribers}{\code{connectapi::Variant$get_subscribers()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="connectapi" data-topic="Variant" data-id="get_url">}\href{../../connectapi/html/Variant.html#method-get_url}{\code{connectapi::Variant$get_url()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="connectapi" data-topic="Variant" data-id="get_url_rev">}\href{../../connectapi/html/Variant.html#method-get_url_rev}{\code{connectapi::Variant$get_url_rev()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="connectapi" data-topic="Variant" data-id="get_variant">}\href{../../connectapi/html/Variant.html#method-get_variant}{\code{connectapi::Variant$get_variant()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="connectapi" data-topic="Variant" data-id="get_variant_remote">}\href{../../connectapi/html/Variant.html#method-get_variant_remote}{\code{connectapi::Variant$get_variant_remote()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="connectapi" data-topic="Variant" data-id="job">}\href{../../connectapi/html/Variant.html#method-job}{\code{connectapi::Variant$job()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="connectapi" data-topic="Variant" data-id="jobs">}\href{../../connectapi/html/Variant.html#method-jobs}{\code{connectapi::Variant$jobs()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="connectapi" data-topic="Variant" data-id="remove_subscriber">}\href{../../connectapi/html/Variant.html#method-remove_subscriber}{\code{connectapi::Variant$remove_subscriber()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="connectapi" data-topic="Variant" data-id="render">}\href{../../connectapi/html/Variant.html#method-render}{\code{connectapi::Variant$render()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="connectapi" data-topic="Variant" data-id="renderings">}\href{../../connectapi/html/Variant.html#method-renderings}{\code{connectapi::Variant$renderings()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="connectapi" data-topic="Variant" data-id="send_mail">}\href{../../connectapi/html/Variant.html#method-send_mail}{\code{connectapi::Variant$send_mail()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="connectapi" data-topic="Variant" data-id="update_variant">}\href{../../connectapi/html/Variant.html#method-update_variant}{\code{connectapi::Variant$update_variant()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VariantSchedule$new(connect, content, key, schedule)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GET"></a>}}
\if{latex}{\out{\hypertarget{method-GET}{}}}
\subsection{Method \code{GET()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VariantSchedule$GET(path)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-POST"></a>}}
\if{latex}{\out{\hypertarget{method-POST}{}}}
\subsection{Method \code{POST()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VariantSchedule$POST(path, body)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DELETE"></a>}}
\if{latex}{\out{\hypertarget{method-DELETE}{}}}
\subsection{Method \code{DELETE()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VariantSchedule$DELETE(path)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_schedule"></a>}}
\if{latex}{\out{\hypertarget{method-set_schedule}{}}}
\subsection{Method \code{set_schedule()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VariantSchedule$set_schedule(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-is_empty"></a>}}
\if{latex}{\out{\hypertarget{method-is_empty}{}}}
\subsection{Method \code{is_empty()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VariantSchedule$is_empty()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VariantSchedule$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_schedule"></a>}}
\if{latex}{\out{\hypertarget{method-get_schedule}{}}}
\subsection{Method \code{get_schedule()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VariantSchedule$get_schedule()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_schedule_remote"></a>}}
\if{latex}{\out{\hypertarget{method-get_schedule_remote}{}}}
\subsection{Method \code{get_schedule_remote()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VariantSchedule$get_schedule_remote()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-describe_schedule"></a>}}
\if{latex}{\out{\hypertarget{method-describe_schedule}{}}}
\subsection{Method \code{describe_schedule()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VariantSchedule$describe_schedule()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VariantSchedule$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
