% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant.R
\name{VariantTask}
\alias{VariantTask}
\title{VariantTask}
\description{
VariantTask

VariantTask
}
\details{
An R6 class that represents a Variant Task
}
\seealso{
Other R6 classes: 
\code{\link{Bundle}},
\code{\link{Content}},
\code{\link{ContentTask}},
\code{\link{Environment}},
\code{\link{PositConnect}},
\code{\link{Task}},
\code{\link{Vanity}},
\code{\link{Variant}},
\code{\link{VariantSchedule}}
}
\concept{R6 classes}
\section{Super classes}{
\code{\link[connectapi:Content]{connectapi::Content}} -> \code{\link[connectapi:Variant]{connectapi::Variant}} -> \code{VariantTask}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{task}}{The task.}

\item{\code{data}}{The variant data.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-VariantTask-new}{\code{VariantTask$new()}}
\item \href{#method-VariantTask-get_task}{\code{VariantTask$get_task()}}
\item \href{#method-VariantTask-add_data}{\code{VariantTask$add_data()}}
\item \href{#method-VariantTask-get_data}{\code{VariantTask$get_data()}}
\item \href{#method-VariantTask-print}{\code{VariantTask$print()}}
\item \href{#method-VariantTask-clone}{\code{VariantTask$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="connectapi" data-topic="Content" data-id="bundle_delete"><a href='../../connectapi/html/Content.html#method-Content-bundle_delete'><code>connectapi::Content$bundle_delete()</code></a></span></li>
<li><span class="pkg-link" data-pkg="connectapi" data-topic="Content" data-id="bundle_download"><a href='../../connectapi/html/Content.html#method-Content-bundle_download'><code>connectapi::Content$bundle_download()</code></a></span></li>
<li><span class="pkg-link" data-pkg="connectapi" data-topic="Content" data-id="danger_delete"><a href='../../connectapi/html/Content.html#method-Content-danger_delete'><code>connectapi::Content$danger_delete()</code></a></span></li>
<li><span class="pkg-link" data-pkg="connectapi" data-topic="Content" data-id="deploy"><a href='../../connectapi/html/Content.html#method-Content-deploy'><code>connectapi::Content$deploy()</code></a></span></li>
<li><span class="pkg-link" data-pkg="connectapi" data-topic="Content" data-id="environment"><a href='../../connectapi/html/Content.html#method-Content-environment'><code>connectapi::Content$environment()</code></a></span></li>
<li><span class="pkg-link" data-pkg="connectapi" data-topic="Content" data-id="environment_all"><a href='../../connectapi/html/Content.html#method-Content-environment_all'><code>connectapi::Content$environment_all()</code></a></span></li>
<li><span class="pkg-link" data-pkg="connectapi" data-topic="Content" data-id="environment_set"><a href='../../connectapi/html/Content.html#method-Content-environment_set'><code>connectapi::Content$environment_set()</code></a></span></li>
<li><span class="pkg-link" data-pkg="connectapi" data-topic="Content" data-id="get_bundles"><a href='../../connectapi/html/Content.html#method-Content-get_bundles'><code>connectapi::Content$get_bundles()</code></a></span></li>
<li><span class="pkg-link" data-pkg="connectapi" data-topic="Content" data-id="get_connect"><a href='../../connectapi/html/Content.html#method-Content-get_connect'><code>connectapi::Content$get_connect()</code></a></span></li>
<li><span class="pkg-link" data-pkg="connectapi" data-topic="Content" data-id="get_content"><a href='../../connectapi/html/Content.html#method-Content-get_content'><code>connectapi::Content$get_content()</code></a></span></li>
<li><span class="pkg-link" data-pkg="connectapi" data-topic="Content" data-id="get_content_remote"><a href='../../connectapi/html/Content.html#method-Content-get_content_remote'><code>connectapi::Content$get_content_remote()</code></a></span></li>
<li><span class="pkg-link" data-pkg="connectapi" data-topic="Content" data-id="get_job"><a href='../../connectapi/html/Content.html#method-Content-get_job'><code>connectapi::Content$get_job()</code></a></span></li>
<li><span class="pkg-link" data-pkg="connectapi" data-topic="Content" data-id="get_jobs"><a href='../../connectapi/html/Content.html#method-Content-get_jobs'><code>connectapi::Content$get_jobs()</code></a></span></li>
<li><span class="pkg-link" data-pkg="connectapi" data-topic="Content" data-id="internal_content"><a href='../../connectapi/html/Content.html#method-Content-internal_content'><code>connectapi::Content$internal_content()</code></a></span></li>
<li><span class="pkg-link" data-pkg="connectapi" data-topic="Content" data-id="permissions"><a href='../../connectapi/html/Content.html#method-Content-permissions'><code>connectapi::Content$permissions()</code></a></span></li>
<li><span class="pkg-link" data-pkg="connectapi" data-topic="Content" data-id="permissions_add"><a href='../../connectapi/html/Content.html#method-Content-permissions_add'><code>connectapi::Content$permissions_add()</code></a></span></li>
<li><span class="pkg-link" data-pkg="connectapi" data-topic="Content" data-id="permissions_delete"><a href='../../connectapi/html/Content.html#method-Content-permissions_delete'><code>connectapi::Content$permissions_delete()</code></a></span></li>
<li><span class="pkg-link" data-pkg="connectapi" data-topic="Content" data-id="permissions_update"><a href='../../connectapi/html/Content.html#method-Content-permissions_update'><code>connectapi::Content$permissions_update()</code></a></span></li>
<li><span class="pkg-link" data-pkg="connectapi" data-topic="Content" data-id="repo_enable"><a href='../../connectapi/html/Content.html#method-Content-repo_enable'><code>connectapi::Content$repo_enable()</code></a></span></li>
<li><span class="pkg-link" data-pkg="connectapi" data-topic="Content" data-id="repo_set"><a href='../../connectapi/html/Content.html#method-Content-repo_set'><code>connectapi::Content$repo_set()</code></a></span></li>
<li><span class="pkg-link" data-pkg="connectapi" data-topic="Content" data-id="runas"><a href='../../connectapi/html/Content.html#method-Content-runas'><code>connectapi::Content$runas()</code></a></span></li>
<li><span class="pkg-link" data-pkg="connectapi" data-topic="Content" data-id="tag_delete"><a href='../../connectapi/html/Content.html#method-Content-tag_delete'><code>connectapi::Content$tag_delete()</code></a></span></li>
<li><span class="pkg-link" data-pkg="connectapi" data-topic="Content" data-id="tag_set"><a href='../../connectapi/html/Content.html#method-Content-tag_set'><code>connectapi::Content$tag_set()</code></a></span></li>
<li><span class="pkg-link" data-pkg="connectapi" data-topic="Content" data-id="tags"><a href='../../connectapi/html/Content.html#method-Content-tags'><code>connectapi::Content$tags()</code></a></span></li>
<li><span class="pkg-link" data-pkg="connectapi" data-topic="Content" data-id="update"><a href='../../connectapi/html/Content.html#method-Content-update'><code>connectapi::Content$update()</code></a></span></li>
<li><span class="pkg-link" data-pkg="connectapi" data-topic="Content" data-id="variants"><a href='../../connectapi/html/Content.html#method-Content-variants'><code>connectapi::Content$variants()</code></a></span></li>
<li><span class="pkg-link" data-pkg="connectapi" data-topic="Variant" data-id="add_subscribers"><a href='../../connectapi/html/Variant.html#method-Variant-add_subscribers'><code>connectapi::Variant$add_subscribers()</code></a></span></li>
<li><span class="pkg-link" data-pkg="connectapi" data-topic="Variant" data-id="get_dashboard_url"><a href='../../connectapi/html/Variant.html#method-Variant-get_dashboard_url'><code>connectapi::Variant$get_dashboard_url()</code></a></span></li>
<li><span class="pkg-link" data-pkg="connectapi" data-topic="Variant" data-id="get_schedule"><a href='../../connectapi/html/Variant.html#method-Variant-get_schedule'><code>connectapi::Variant$get_schedule()</code></a></span></li>
<li><span class="pkg-link" data-pkg="connectapi" data-topic="Variant" data-id="get_schedule_remote"><a href='../../connectapi/html/Variant.html#method-Variant-get_schedule_remote'><code>connectapi::Variant$get_schedule_remote()</code></a></span></li>
<li><span class="pkg-link" data-pkg="connectapi" data-topic="Variant" data-id="get_subscribers"><a href='../../connectapi/html/Variant.html#method-Variant-get_subscribers'><code>connectapi::Variant$get_subscribers()</code></a></span></li>
<li><span class="pkg-link" data-pkg="connectapi" data-topic="Variant" data-id="get_url"><a href='../../connectapi/html/Variant.html#method-Variant-get_url'><code>connectapi::Variant$get_url()</code></a></span></li>
<li><span class="pkg-link" data-pkg="connectapi" data-topic="Variant" data-id="get_url_rev"><a href='../../connectapi/html/Variant.html#method-Variant-get_url_rev'><code>connectapi::Variant$get_url_rev()</code></a></span></li>
<li><span class="pkg-link" data-pkg="connectapi" data-topic="Variant" data-id="get_variant"><a href='../../connectapi/html/Variant.html#method-Variant-get_variant'><code>connectapi::Variant$get_variant()</code></a></span></li>
<li><span class="pkg-link" data-pkg="connectapi" data-topic="Variant" data-id="get_variant_remote"><a href='../../connectapi/html/Variant.html#method-Variant-get_variant_remote'><code>connectapi::Variant$get_variant_remote()</code></a></span></li>
<li><span class="pkg-link" data-pkg="connectapi" data-topic="Variant" data-id="job"><a href='../../connectapi/html/Variant.html#method-Variant-job'><code>connectapi::Variant$job()</code></a></span></li>
<li><span class="pkg-link" data-pkg="connectapi" data-topic="Variant" data-id="jobs"><a href='../../connectapi/html/Variant.html#method-Variant-jobs'><code>connectapi::Variant$jobs()</code></a></span></li>
<li><span class="pkg-link" data-pkg="connectapi" data-topic="Variant" data-id="remove_subscriber"><a href='../../connectapi/html/Variant.html#method-Variant-remove_subscriber'><code>connectapi::Variant$remove_subscriber()</code></a></span></li>
<li><span class="pkg-link" data-pkg="connectapi" data-topic="Variant" data-id="render"><a href='../../connectapi/html/Variant.html#method-Variant-render'><code>connectapi::Variant$render()</code></a></span></li>
<li><span class="pkg-link" data-pkg="connectapi" data-topic="Variant" data-id="renderings"><a href='../../connectapi/html/Variant.html#method-Variant-renderings'><code>connectapi::Variant$renderings()</code></a></span></li>
<li><span class="pkg-link" data-pkg="connectapi" data-topic="Variant" data-id="send_mail"><a href='../../connectapi/html/Variant.html#method-Variant-send_mail'><code>connectapi::Variant$send_mail()</code></a></span></li>
<li><span class="pkg-link" data-pkg="connectapi" data-topic="Variant" data-id="update_variant"><a href='../../connectapi/html/Variant.html#method-Variant-update_variant'><code>connectapi::Variant$update_variant()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VariantTask-new"></a>}}
\if{latex}{\out{\hypertarget{method-VariantTask-new}{}}}
\subsection{Method \code{new()}}{
Initialize this variant task.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VariantTask$new(connect, content, key, task)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{connect}}{The \code{Connect} instance.}

\item{\code{content}}{The \code{Content} instance.}

\item{\code{key}}{The variant key.}

\item{\code{task}}{The task data.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VariantTask-get_task"></a>}}
\if{latex}{\out{\hypertarget{method-VariantTask-get_task}{}}}
\subsection{Method \code{get_task()}}{
Return the underlying task.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VariantTask$get_task()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VariantTask-add_data"></a>}}
\if{latex}{\out{\hypertarget{method-VariantTask-add_data}{}}}
\subsection{Method \code{add_data()}}{
Set the data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VariantTask$add_data(data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{The data.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VariantTask-get_data"></a>}}
\if{latex}{\out{\hypertarget{method-VariantTask-get_data}{}}}
\subsection{Method \code{get_data()}}{
Get the data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VariantTask$get_data()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VariantTask-print"></a>}}
\if{latex}{\out{\hypertarget{method-VariantTask-print}{}}}
\subsection{Method \code{print()}}{
Print this object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VariantTask$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Unused.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VariantTask-clone"></a>}}
\if{latex}{\out{\hypertarget{method-VariantTask-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VariantTask$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
