% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant.R
\name{VariantTask}
\alias{VariantTask}
\title{VariantTask}
\description{
An R6 class that represents a Variant Task
}
\seealso{
Other R6 classes: 
\code{\link{Bundle}},
\code{\link{ContentTask}},
\code{\link{Content}},
\code{\link{Environment}},
\code{\link{RStudioConnect}},
\code{\link{Task}},
\code{\link{Vanity}},
\code{\link{VariantSchedule}},
\code{\link{Variant}}
}
\concept{R6 classes}
\section{Super classes}{
\code{\link[connectapi:Content]{connectapi::Content}} -> \code{\link[connectapi:Variant]{connectapi::Variant}} -> \code{VariantTask}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{VariantTask$new()}}
\item \href{#method-get_task}{\code{VariantTask$get_task()}}
\item \href{#method-add_data}{\code{VariantTask$add_data()}}
\item \href{#method-get_data}{\code{VariantTask$get_data()}}
\item \href{#method-print}{\code{VariantTask$print()}}
\item \href{#method-clone}{\code{VariantTask$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="connectapi" data-topic="Content" data-id="bundle_delete">}\href{../../connectapi/html/Content.html#method-bundle_delete}{\code{connectapi::Content$bundle_delete()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="connectapi" data-topic="Content" data-id="bundle_download">}\href{../../connectapi/html/Content.html#method-bundle_download}{\code{connectapi::Content$bundle_download()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="connectapi" data-topic="Content" data-id="danger_delete">}\href{../../connectapi/html/Content.html#method-danger_delete}{\code{connectapi::Content$danger_delete()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="connectapi" data-topic="Content" data-id="deploy">}\href{../../connectapi/html/Content.html#method-deploy}{\code{connectapi::Content$deploy()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="connectapi" data-topic="Content" data-id="environment">}\href{../../connectapi/html/Content.html#method-environment}{\code{connectapi::Content$environment()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="connectapi" data-topic="Content" data-id="environment_all">}\href{../../connectapi/html/Content.html#method-environment_all}{\code{connectapi::Content$environment_all()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="connectapi" data-topic="Content" data-id="environment_set">}\href{../../connectapi/html/Content.html#method-environment_set}{\code{connectapi::Content$environment_set()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="connectapi" data-topic="Content" data-id="get_bundles">}\href{../../connectapi/html/Content.html#method-get_bundles}{\code{connectapi::Content$get_bundles()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="connectapi" data-topic="Content" data-id="get_connect">}\href{../../connectapi/html/Content.html#method-get_connect}{\code{connectapi::Content$get_connect()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="connectapi" data-topic="Content" data-id="get_content">}\href{../../connectapi/html/Content.html#method-get_content}{\code{connectapi::Content$get_content()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="connectapi" data-topic="Content" data-id="get_content_remote">}\href{../../connectapi/html/Content.html#method-get_content_remote}{\code{connectapi::Content$get_content_remote()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="connectapi" data-topic="Content" data-id="get_job">}\href{../../connectapi/html/Content.html#method-get_job}{\code{connectapi::Content$get_job()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="connectapi" data-topic="Content" data-id="get_jobs">}\href{../../connectapi/html/Content.html#method-get_jobs}{\code{connectapi::Content$get_jobs()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="connectapi" data-topic="Content" data-id="internal_content">}\href{../../connectapi/html/Content.html#method-internal_content}{\code{connectapi::Content$internal_content()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="connectapi" data-topic="Content" data-id="permissions">}\href{../../connectapi/html/Content.html#method-permissions}{\code{connectapi::Content$permissions()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="connectapi" data-topic="Content" data-id="permissions_add">}\href{../../connectapi/html/Content.html#method-permissions_add}{\code{connectapi::Content$permissions_add()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="connectapi" data-topic="Content" data-id="permissions_delete">}\href{../../connectapi/html/Content.html#method-permissions_delete}{\code{connectapi::Content$permissions_delete()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="connectapi" data-topic="Content" data-id="permissions_update">}\href{../../connectapi/html/Content.html#method-permissions_update}{\code{connectapi::Content$permissions_update()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="connectapi" data-topic="Content" data-id="repo_enable">}\href{../../connectapi/html/Content.html#method-repo_enable}{\code{connectapi::Content$repo_enable()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="connectapi" data-topic="Content" data-id="repo_set">}\href{../../connectapi/html/Content.html#method-repo_set}{\code{connectapi::Content$repo_set()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="connectapi" data-topic="Content" data-id="runas">}\href{../../connectapi/html/Content.html#method-runas}{\code{connectapi::Content$runas()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="connectapi" data-topic="Content" data-id="tag_delete">}\href{../../connectapi/html/Content.html#method-tag_delete}{\code{connectapi::Content$tag_delete()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="connectapi" data-topic="Content" data-id="tag_set">}\href{../../connectapi/html/Content.html#method-tag_set}{\code{connectapi::Content$tag_set()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="connectapi" data-topic="Content" data-id="tags">}\href{../../connectapi/html/Content.html#method-tags}{\code{connectapi::Content$tags()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="connectapi" data-topic="Content" data-id="update">}\href{../../connectapi/html/Content.html#method-update}{\code{connectapi::Content$update()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="connectapi" data-topic="Content" data-id="variants">}\href{../../connectapi/html/Content.html#method-variants}{\code{connectapi::Content$variants()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="connectapi" data-topic="Variant" data-id="add_subscribers">}\href{../../connectapi/html/Variant.html#method-add_subscribers}{\code{connectapi::Variant$add_subscribers()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="connectapi" data-topic="Variant" data-id="get_dashboard_url">}\href{../../connectapi/html/Variant.html#method-get_dashboard_url}{\code{connectapi::Variant$get_dashboard_url()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="connectapi" data-topic="Variant" data-id="get_schedule">}\href{../../connectapi/html/Variant.html#method-get_schedule}{\code{connectapi::Variant$get_schedule()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="connectapi" data-topic="Variant" data-id="get_schedule_remote">}\href{../../connectapi/html/Variant.html#method-get_schedule_remote}{\code{connectapi::Variant$get_schedule_remote()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="connectapi" data-topic="Variant" data-id="get_subscribers">}\href{../../connectapi/html/Variant.html#method-get_subscribers}{\code{connectapi::Variant$get_subscribers()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="connectapi" data-topic="Variant" data-id="get_url">}\href{../../connectapi/html/Variant.html#method-get_url}{\code{connectapi::Variant$get_url()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="connectapi" data-topic="Variant" data-id="get_url_rev">}\href{../../connectapi/html/Variant.html#method-get_url_rev}{\code{connectapi::Variant$get_url_rev()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="connectapi" data-topic="Variant" data-id="get_variant">}\href{../../connectapi/html/Variant.html#method-get_variant}{\code{connectapi::Variant$get_variant()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="connectapi" data-topic="Variant" data-id="get_variant_remote">}\href{../../connectapi/html/Variant.html#method-get_variant_remote}{\code{connectapi::Variant$get_variant_remote()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="connectapi" data-topic="Variant" data-id="job">}\href{../../connectapi/html/Variant.html#method-job}{\code{connectapi::Variant$job()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="connectapi" data-topic="Variant" data-id="jobs">}\href{../../connectapi/html/Variant.html#method-jobs}{\code{connectapi::Variant$jobs()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="connectapi" data-topic="Variant" data-id="remove_subscriber">}\href{../../connectapi/html/Variant.html#method-remove_subscriber}{\code{connectapi::Variant$remove_subscriber()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="connectapi" data-topic="Variant" data-id="render">}\href{../../connectapi/html/Variant.html#method-render}{\code{connectapi::Variant$render()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="connectapi" data-topic="Variant" data-id="renderings">}\href{../../connectapi/html/Variant.html#method-renderings}{\code{connectapi::Variant$renderings()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="connectapi" data-topic="Variant" data-id="send_mail">}\href{../../connectapi/html/Variant.html#method-send_mail}{\code{connectapi::Variant$send_mail()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="connectapi" data-topic="Variant" data-id="update_variant">}\href{../../connectapi/html/Variant.html#method-update_variant}{\code{connectapi::Variant$update_variant()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VariantTask$new(connect, content, key, task)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_task"></a>}}
\if{latex}{\out{\hypertarget{method-get_task}{}}}
\subsection{Method \code{get_task()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VariantTask$get_task()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_data"></a>}}
\if{latex}{\out{\hypertarget{method-add_data}{}}}
\subsection{Method \code{add_data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VariantTask$add_data(data)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_data"></a>}}
\if{latex}{\out{\hypertarget{method-get_data}{}}}
\subsection{Method \code{get_data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VariantTask$get_data()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VariantTask$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VariantTask$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
