% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content.R
\name{verify_content_name}
\alias{verify_content_name}
\title{Verify Content Name}
\usage{
verify_content_name(name)
}
\arguments{
\item{name}{The proposed content name}
}
\value{
The name (or an error if invalid)
}
\description{
Ensures that a content name fits the specifications / requirements of RStudio
Connect. Throws an error if content name is invalid. Content names (as of the
time of writing) must be between 3 and 64 alphanumeric characters, dashes,
and underscores
}
\seealso{
connectapi::create_random_name

Other content functions: 
\code{\link{acl_add_user}()},
\code{\link{content_item}()},
\code{\link{content_title}()},
\code{\link{create_random_name}()},
\code{\link{dashboard_url_chr}()},
\code{\link{dashboard_url}()},
\code{\link{delete_vanity_url}()},
\code{\link{deploy_repo}()},
\code{\link{get_acl_user}()},
\code{\link{get_bundles}()},
\code{\link{get_environment}()},
\code{\link{get_image}()},
\code{\link{get_jobs}()},
\code{\link{get_vanity_url}()},
\code{\link{git}},
\code{\link{permissions}},
\code{\link{set_image_path}()},
\code{\link{set_run_as}()},
\code{\link{set_vanity_url}()},
\code{\link{swap_vanity_url}()}
}
\concept{content functions}
